% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_matrixes.R
\name{AppendReadCounts}
\alias{AppendReadCounts}
\title{append read counts for additional conditions}
\usage{
AppendReadCounts(cnt, aln, grg, maxgap = 0L, minoverlap = 1L,
  type = "any", ignore.strand = T, ...)
}
\arguments{
\item{cnt}{a numeric matrix representing read counts and where
rows = observations and columns = measurement conditions.}

\item{aln}{a list with named elements, containing exclusively bam file paths or
\link{GAlignments} objects.}

\item{grg}{a \link{GRanges} object defining the genomic intervals of interest.}

\item{maxgap}{
    See \code{\link[IRanges]{findOverlaps}} in the \pkg{IRanges} package
    for a description of these arguments.
  }

\item{minoverlap}{
    See \code{\link[IRanges]{findOverlaps}} in the \pkg{IRanges} package
    for a description of these arguments.
  }

\item{type}{
    See \code{\link[IRanges]{findOverlaps}} in the \pkg{IRanges} package
    for a description of these arguments.
  }

\item{ignore.strand}{
    When set to \code{TRUE}, the strand information is ignored in the
    overlap calculations.
  }

\item{...}{additional arguments passed to the \link{readGAlignments} function.}
}
\value{
functions \code{ReadCountMatrix}, \code{AppendReadCounts}, and
\code{UpdateReadCounts} return a read counts matrix where
rows = observations and columns = measurement conditions.
}
\description{
append read counts for additional conditions
}
\seealso{
\link{LoadMappedReads},
\link{UpdateReadCounts},
\link{ReadCountMatrix},
\link{MakeReadCounts}
}
\keyword{internal}
