% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RData.R
\name{AvailableObj}
\alias{AvailableObj}
\title{Check if an R object is available as RData}
\usage{
AvailableObj(obj, path = NULL, name = NULL)
}
\arguments{
\item{obj}{an R object previously saved using \link{SaveObj}.}

\item{path}{folder location used to load the R object.
When none is specified it is the current working directory.}

\item{name}{a file name. When omitted (default) this name is automatically set to the
name of the R object being passed as argument.}
}
\value{
logical, \code{TRUE} when RData is available and \code{FALSE} otherwise.
}
\description{
Check if an R object is available as RData
}
\seealso{
\link{LoadObj},
  \link{SaveObj}
}
\keyword{internal}
