% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invariability_DEPRECATED.R
\name{bg_grp}
\alias{bg_grp}
\title{background population}
\usage{
bg_grp(x, grp, bg_cols)
}
\arguments{
\item{x}{}

\item{grp}{}

\item{bg_cols}{}
}
\value{
cluster id
}
\description{
background population
}
