% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRD.R
\name{bg_ranking}
\alias{bg_ranking}
\title{Average ranking for control measurements}
\usage{
bg_ranking(x, by, controls)
}
\arguments{
\item{x}{matrix of read counts
(rows = observations, columns = measurement conditions).}

\item{by}{grouping index.}

\item{controls}{columns corresponding to control measurements (Input, IgG, etc.)}
}
\value{
group identifiers ordered by decreasing average difference
between controls and other measurement conditions.
}
\description{
Average ranking for control measurements
}
\seealso{
\link{BRD}
}
\keyword{internal}
