% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRD.R
\name{BRD}
\alias{BRD}
\title{Background read density clustering}
\usage{
BRD(cnt, controls = NULL, dither = 3, smobs = T, cvt = 0.5, npc = NA,
  zscore = T, knn = 200, cst = 0.5, ncl = 3, mcs = 100,
  progress = F)
}
\arguments{
\item{cnt}{matrix of read counts
(rows = observations, columns = measurement conditions).}

\item{controls}{columns corresponding to control measurements (Input, IgG, etc.)}

\item{dither}{number of replicates for count dithering: 0 = no dithering,
1 = single (default), 2 = duplicate, 3 = triplicate, etc.}

\item{smobs}{subtract the mean count value of each observation (logical, default = F).}

\item{cvt}{percentage of variance in principal components to be ignored.
The value of \code{cvt} determines the stringency of the dimensionality
reduction which is performed by Principal Component Analysis.
Dimensions retained after projection of the read count dataset on its
principal components will preserve \code{100 * (1 - cvt)}% of the total
variance.
For instance with the default value (\code{cvt} = 0.05),
95% of the total variance will remain after dimensionality reduction.
Setting cvt = 0 bypasses the dimensionality reduction and performs
density estimations in the original multivariate measurement space
of the dataset.}

\item{npc}{number of dimensions retained after projection on principal components.
When \code{npc} is specified, the \code{cvt} parameter is ignored.}

\item{zscore}{transform counts into z-scores (logical, default = T).}

\item{knn}{number of nearest neighbors, which corresponds to the smoothing parameter
of estimated densities (larger values = smoother).}

\item{cst}{core subset density threshold}

\item{ncl}{number of clusters}

\item{mcs}{minimum size of the background cluster}

\item{progress}{show progress bar (logical, default = F).}
}
\value{
BRD returns a \code{list}
}
\description{
Background read density clustering
}
