% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildGeneFeatures.R
\name{BuildGeneFeatures}
\alias{BuildGeneFeatures}
\title{Build an R environment containing genomic features from Ensembl}
\usage{
BuildGeneFeatures(organism, gene_id = "ensembl_gene_id",
  gene_name = "hgnc_symbol")
}
\arguments{
\item{organism}{list.}

\item{gene_id}{character (default = "ensembl_gene_id").}

\item{gene_name}{character (default = "hgnc_symbol").}
}
\value{
\code{BuildGeneFeatures} returns an \link{environment}.
}
\description{
Build an R environment containing genomic features from Ensembl
}
