% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCSC.R
\name{BuildSeqInfo}
\alias{BuildSeqInfo}
\title{Retrieve chromosome information from UCSC and build Seqinfo object}
\usage{
BuildSeqInfo(genome)
}
\arguments{
\item{genome}{name of an UCSC genome build.}
}
\value{
\code{BuildSeqInfo} returns a \link{Seqinfo} object.
}
\description{
Retrieve chromosome information from UCSC and build Seqinfo object
}
