% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDaDaDR_2D.R
\name{CDaDaDR.2D}
\alias{CDaDaDR.2D}
\title{Density estimation in reduced dimensions}
\usage{
CDaDaDR.2D(cnt, movs = NULL, bins = NULL, smoothing = NULL,
  dither = NULL, zscore = NULL, method = c("pca", "ica"))
}
\arguments{
\item{cnt}{matrix of read counts
(rows = observations, columns = measurement conditions).}

\item{movs}{precomputed mean count values to be subtracted to each observation
(default = NULL, zero).}

\item{bins}{number of bins per principal component for density estimations
(default = 500).}

\item{smoothing}{number of consecutive bins for local average smoothing of estimated densities
(default = 10).}

\item{dither}{number of replicates of the count dithering performed by \link{DitherCounts}:
1 = single, 2 = duplicate, 3 = triplicate, etc. (default = 5).}

\item{zscore}{transform read count projections into z-scores (default = TRUE, recommended).}

\item{method}{projection method, either "pca" (default) performed by \link{prcomp},
or "ica" performed by \link{icafast}.}
}
\value{
\code{CDaDaDR.2D} returns a list with the following elements:
\item{parameters}{list with the value of each CDaDaDR.2D argument}
\item{status}{list with informations on the dimensionality reduction}
\item{density}{estimated density at each observation}
\item{projection}{transformed count matrix}
}
\description{
Count Density after Dithering and Dimensionality Reduction (CDaDaDR).
}
\seealso{
\link{BRD},
  \link{PlotBRD},
  \link{DitherCounts},
  \link{Barbouille::ASH2D}
}
\keyword{internal}
