% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdadadr.R
\name{CDaDaDR}
\alias{CDaDaDR}
\title{Count Density after Dithering and Dimensionality Reduction}
\usage{
CDaDaDR(cnt, knn, smobs = F, movs = NULL, dither = 1, npc = NA,
  zscore = T, rare = 0.01, method = c("pca", "ica"), progress = F)
}
\arguments{
\item{cnt}{matrix of read counts
(rows = observations, columns = measurement conditions).}

\item{knn}{number of nearest neighbors, which corresponds to the smoothing parameter
of estimated densities (larger values = smoother).}

\item{smobs}{subtract the mean count value of each observation (logical, default = F).}

\item{movs}{precomputed mean count values to be subtracted to each observation
(default = NULL, zero).}

\item{dither}{number of replicates for count dithering:
1 = single (default), 2 = duplicate, 3 = triplicate, etc.}

\item{npc}{number of dimensions retained after projection on principal components.
When \code{npc} is specified, the \code{cvt} parameter is ignored.}

\item{zscore}{transform counts into z-scores (logical, default = T).}

\item{rare}{defines exceptionally low density values (default = 0.01).}

\item{method}{projection method, either "pca" (default) or "ica".}

\item{progress}{show progress bar (logical, default = F).}
}
\value{
CDaDaDR returns a \code{list} with the following elements:
\item{parameters}{list with the value of each CDaDaDR argument}
\item{status}{list with informations on the dimensionality reduction}
\item{density}{knn density}
\item{projection}{transformed count matrix}
}
\description{
Count Density after Dithering and Dimensionality Reduction
}
\seealso{
\link{BRD},
  \link{DitherCounts},
  \link{knn_density}
}
\keyword{internal}
