% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{centerOfDifferences}
\alias{centerOfDifferences}
\title{Center of differences between log transformed counts in different samples}
\usage{
centerOfDifferences(x, invariable = NULL, center = NULL, replicates = 0,
  plot = F, xylim = NULL, ...)
}
\arguments{
\item{x}{matrix of log transformed read counts (columns = samples, rows = genomic intervals)}

\item{invariable}{logical vector defining the subset of rows assumed to be invariable (default = all)}

\item{center}{function for center of difference calculations.
The default center function is \link{findMode} which identifies the mode in
a kernel density estimation of the distribution of read count differences
using the generic R function \code{density}.
This default function can be replaced by \code{median}, \code{mean}, or
similar functions.}

\item{replicates}{logical matrix indicating replicate columns (default = none)}

\item{plot}{logical, activates the drawing of control scatterplots (default = F)}

\item{xylim}{defining ranges of x and y axes for scatterplots (default = auto)}

\item{...}{graphical parameters passed to the \link{scPlotCounts} function.}
}
\value{
matrix of pairwise center of differences between all samples
}
\description{
Center of differences between log transformed counts in different samples
}

