% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_data.R
\docType{data}
\name{CGI_hg38}
\alias{CGI_hg38}
\title{UCSC CGIs - Homo sapiens - GRCh38 / hg38}
\format{An object of class \code{GRanges} of length 30477.}
\usage{
CGI_hg38
}
\description{
UCSC CGIs - Homo sapiens - GRCh38 / hg38
}
\keyword{datasets}
