% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_data.R
\docType{data}
\name{CGI_mm10}
\alias{CGI_mm10}
\title{UCSC CGIs - Mus musculus - GRCm38 / mm10}
\format{An object of class \code{GRanges} of length 16023.}
\usage{
CGI_mm10
}
\description{
UCSC CGIs - Mus musculus - GRCm38 / mm10
}
\keyword{datasets}
