% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{CleanCounts}
\alias{CleanCounts}
\title{Remove observations with missing values}
\usage{
CleanCounts(cnt)
}
\arguments{
\item{cnt}{matrix of read counts
(rows = observations, columns = measurement conditions).}
}
\value{
matrix
}
\description{
Remove observations with missing values
}
\seealso{
\link{FiniteValues}
  \link{DetectCounts}
}
\keyword{internal}
