% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicRanges.R
\name{CleanupGRanges}
\alias{CleanupGRanges}
\title{Cleanup genomic ranges}
\usage{
CleanupGRanges(grg, seqinfo, organism, blacklist = NULL, keep.strand = F)
}
\arguments{
\item{grg}{a \link{GRanges} object.}

\item{seqinfo}{a \link{Seqinfo} object.}

\item{organism}{character.}

\item{blacklist}{a \link{GRanges} object (default = NULL, none).}

\item{keep.strand}{logical (default = F, no).}
}
\value{
\code{CleanupGRanges} returns a \link{GRanges} object.
}
\description{
Cleanup genomic ranges (e.g. for imported ChIP-seq peaks from bed files).
}
\keyword{internal}
