% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{ClonalGaussian}
\alias{ClonalGaussian}
\title{ClonalGaussian}
\usage{
ClonalGaussian(n, mu, sigma)
}
\arguments{
\item{n}{number}

\item{mu}{mean}

\item{sigma}{strandard deviation}
}
\value{
list
}
\description{
ClonalGaussian
}
\keyword{internal}
