% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{ClonalUniform}
\alias{ClonalUniform}
\title{ClonalUniform}
\usage{
ClonalUniform(n, mu, delta)
}
\arguments{
\item{n}{number}

\item{mu}{mean}

\item{delta}{half width of the range}
}
\value{
list
}
\description{
ClonalUniform
}
\keyword{internal}
