% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{correctByVarianceWeights}
\alias{correctByVarianceWeights}
\title{variance stabilization on log-transformed counts using voom-like weights}
\usage{
correctByVarianceWeights(cnt, refMean, refWeights)
}
\arguments{
\item{\code{cnt}}{matrix of read counts (columns = samples, rows = intervals)}

\item{\code{refWeights}}{}
}
\value{
\code{matrix}
}
\description{
variance stabilization on log-transformed counts using voom-like weights
}

