% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicRanges.R
\name{CoveredLength}
\alias{CoveredLength}
\title{Compute the genomic length covered by sequencing reads}
\usage{
CoveredLength(cvg)
}
\arguments{
\item{cvg}{a coverage object, see \link{GenomicAlignments::coverage}.}
}
\value{
\code{CoveredLength} returns the genomic length covered by sequencing reads.
}
\description{
Compute the genomic length covered by sequencing reads
}
\keyword{internal}
