% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{CovMat2D}
\alias{CovMat2D}
\title{CovMat2D}
\usage{
CovMat2D(alpha, sx, sy)
}
\arguments{
\item{alpha}{angle in radians}

\item{sx}{sigma2(x)}

\item{sy}{sigma2(y)}
}
\value{
covariance matrix
}
\description{
CovMat2D
}
\keyword{internal}
