% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPSES.R
\name{CPSES}
\alias{CPSES}
\title{Combined Pairwise Signal Extraction Scaling (SES)}
\usage{
CPSES(cnt, dither = 1)
}
\arguments{
\item{dither}{number of replicates for count dithering:
1 = single (default), 2 = duplicate, 3 = triplicate, etc.}

\item{x}{matrix of read counts
(rows = observations, columns = measurement conditions).}
}
\value{
\code{CPSES} returns a vector of scaling factors.
}
\description{
Combined Pairwise Signal Extraction Scaling (SES)
}
\seealso{
\link{BRD},
\link{Diaz.SES}
}
\keyword{internal}
