% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_simulation.R
\name{DefineSimulation}
\alias{DefineSimulation}
\title{Define different populations for ChIP-seq simulations}
\usage{
DefineSimulation(chip = 3, replicate = 1, input = 2,
  enrichment = c(1, 3), patterns = "^")
}
\arguments{
\item{chip}{number of ChIP profiles (default = 3).}

\item{replicate}{number of replicates per ChIP profile (default = 1, no replicate).}

\item{input}{number of Input profiles (default = 2).}

\item{enrichment}{numeric vector defining the range of ChIP enrichements (default = c(1, 3)).}

\item{patterns}{character vector, made only of symbols "+", "-", "^" and "v".}
}
\value{
\code{DefineSimulation} returns a \code{matrix} of parameters.
}
\description{
Define different populations for ChIP-seq simulations
}
\seealso{
\link{MakeSimulation}
}
\keyword{internal}
