% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{densityByKNN}
\alias{densityByKNN}
\title{Use k-nearest neighbors as an estimator of the density P(x) ~ k / N V}
\usage{
densityByKNN(qry, k, sbj = NULL, FUN = NULL)
}
\arguments{
\item{FUN}{}
}
\value{
\code{list}
}
\description{
Use k-nearest neighbors as an estimator of the density P(x) ~ k / N V
}

