% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRD.R
\name{DensityCorrectedByIntensity}
\alias{DensityCorrectedByIntensity}
\title{DensityCorrectedByIntensity}
\usage{
DensityCorrectedByIntensity(d, i, k)
}
\arguments{
\item{d}{density vector}

\item{i}{intensity vector}
}
\value{
\code{codeDensityCorrectedByIntensity} returns the vector of corrected
densities.
}
\description{
DensityCorrectedByIntensity
}
\keyword{internal}
