% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{DetectCounts}
\alias{DetectCounts}
\title{DetectCounts}
\usage{
DetectCounts(cnt, detailed = F)
}
\arguments{
\item{cnt}{matrix of read counts
(rows = observations, columns = measurement conditions).}

\item{detailed}{logical, if TRUE return min and max read counts in each interval}
}
\value{
\code{DetectCounts} returns a data.frame with the following columns:
\item{all}{logical vector indicating rows of the \code{cnt} matrix with at least one count in all columns}
\item{none}{logical vector indicating rows of the \code{cnt} matrix with no counts in all columns}
\item{nbr}{integer vector representing the number of columns with at least one counts for each row of the \code{cnt} matrix}
If \code{detailed} is equal to \code{TRUE}, the returned \code{data.frame} also includes the following columns:
\item{min}{integer vector representing the maximum count value for each row of the \code{cnt} matrix}
\item{max}{integer vector representing the minimum count value for each row of the \code{cnt} matrix}
}
\description{
DetectCounts
}
\seealso{
\link{NonZeroCounts},
  \link{ReadCountMatrix}
}
\keyword{internal}
