% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicRanges.R
\name{df2grg}
\alias{df2grg}
\title{Convert \code{data.frame} to \link{GRanges}}
\usage{
df2grg(df, chr = 1, start = 2, end = 3, strand = NULL,
  xidx = NULL, xlbl = NULL)
}
\arguments{
\item{df}{a data.frame defining genomic ranges.}

\item{chr}{column for chromsome identifiers (default = 1).}

\item{start}{column for start position (default = 2).}

\item{end}{column for end position (default = 3).}

\item{strand}{column for strand (default = NULL, e.g. none).}

\item{xidx}{list of extra column indexes to be imported.
By default all available columns are imported.}

\item{xlbl}{character vector defining the name of imported extra columns
(default = NULL).}
}
\value{
\code{df2grg} returns a \link{GRanges} object.
}
\description{
Convert \code{data.frame} to \link{GRanges}
}
\seealso{
\link{ImportGenomicRanges}
}
\keyword{internal}
