% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPSES.R
\name{Diaz.SES}
\alias{Diaz.SES}
\title{Signal Extraction Scaling (Diaz et al., 2012)}
\usage{
Diaz.SES(x, y)
}
\arguments{
\item{x}{Input or IgG raw read counts}

\item{y}{ChIP raw read counts}
}
\value{
list
}
\description{
Signal Extraction Scaling (Diaz et al., 2012)
}
