% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{ditherCounts}
\alias{ditherCounts}
\title{Dither read counts}
\usage{
ditherCounts(cnt, pdf = c("triangle", "normal"), fwhm = 1, min.cnt = 0.5,
  missing.cnt = F)
}
\arguments{
\item{\code{cnt}}{matrix of read counts (columns = samples, rows = intervals)}

\item{\code{pdf}}{probability density of dithering values, either "triangle" (default) or "normal"}

\item{\code{fwhm}}{full width at half maximum for normal pdf, default = 1}

\item{\code{min.cnt}}{minimum count value after dithering, default = 0.5}

\item{\code{missing.cnt}}{logical, if TRUE generate random values for a fraction of null counts}
}
\value{
\code{matrix}
}
\description{
Dither read counts
}

