% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{dualHist}
\alias{dualHist}
\title{Histogram}
\usage{
dualHist(x, y, breaks = 100, x.name = NULL, y.name = NULL,
  x.col = rgb(0, 0, 0, 0.5), y.col = rgb(1, 0, 0, 0.5), ...)
}
\value{
produce an histogram in the active graphic device
}
\description{
Histogram
}

