% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_data.R
\docType{data}
\name{EGA91_mouse}
\alias{EGA91_mouse}
\title{Ensembl genes - Mus musculus - GRCm38 / mm10 (rel. 91, Dec. 2017)}
\format{An object of class \code{environment} of length 12.}
\usage{
EGA91_mouse
}
\description{
Ensembl genes - Mus musculus - GRCm38 / mm10 (rel. 91, Dec. 2017)
}
\keyword{datasets}
