% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{empty.plot}
\alias{empty.plot}
\title{empty.plot}
\usage{
empty.plot(axes = T, xlab = "", ylab = "", ...)
}
\arguments{
\item{axes}{logical}

\item{xlab}{character}

\item{ylab}{character}

\item{...}{}
}
\description{
empty.plot
}
\keyword{internal}
