% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{estimateVarianceWeights}
\alias{estimateVarianceWeights}
\title{voom-like variance weights on log-transformed counts}
\usage{
estimateVarianceWeights(cnt, span = 0.5, plot = F)
}
\arguments{
\item{\code{cnt}}{matrix of read counts (columns = samples, rows = intervals)}

\item{\code{plot}}{}
}
\value{
\code{list}
}
\description{
voom-like variance weights on log-transformed counts
}

