% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{ExtractColumns}
\alias{ExtractColumns}
\title{Extract count data}
\usage{
ExtractColumns(x, lst)
}
\arguments{
\item{x}{list of read count matrices
(rows = observations, columns = measurement conditions).}

\item{lst}{column names or indices to be extracted}
}
\value{
\code{ExtractColumns} returns a list.
}
\description{
Extract count data
}
\seealso{
\link{JoinColumns},
\link{MakeReadCountMatrix},
\link{makeReadCountsInROI}
}
\keyword{internal}
