% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{findMode}
\alias{findMode}
\title{Naive mode estimation}
\usage{
findMode(x, na.rm)
}
\arguments{
\item{\code{x}}{\code{vector}}

\item{\code{na.rm}}{\code{logical}}
}
\value{
scalar
}
\description{
Naive mode estimation
}

