% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{finiteValues}
\alias{finiteValues}
\title{Find finite values consistently in all samples (log transformed counts)}
\usage{
finiteValues(x)
}
\arguments{
\item{x}{numeric matrix}
}
\value{
finiteValues returns a \code{logical} vector
}
\description{
Find finite values consistently in all samples (log transformed counts)
}
