% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.test.R
\name{fitNormal}
\alias{fitNormal}
\title{Fit normal distribution to the data and compute associated p-values}
\usage{
fitNormal(m, invariant = NULL, txt = "", plot = F, brks.len = 70)
}
\arguments{
\item{plot}{}
}
\value{
\code{list}
}
\description{
Fit normal distribution to the data and compute associated p-values
}

