% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicRanges.R
\name{FlatGRanges}
\alias{FlatGRanges}
\title{Reduce genomic intervals}
\usage{
FlatGRanges(x, seqinfo)
}
\arguments{
\item{x}{a \link{GRanges} object.}

\item{seqinfo}{a \link{Seqinfo} object.}
}
\value{
\code{FlatGRanges} returns a \link{GRanges} object.
}
\description{
Reduce genomic intervals
}
\keyword{internal}
