% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fmap}
\alias{fmap}
\title{Alias for mapply with USE.NAMES = F}
\usage{
fmap(...)
}
\arguments{
\item{...}{see \link{mapply}.}
}
\value{
see \link{mapply}.
}
\description{
Alias for mapply with USE.NAMES = F
}
\keyword{internal}
