% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicRanges.R
\name{GenomicCoverage}
\alias{GenomicCoverage}
\title{Total coverage of genomic intervals}
\usage{
GenomicCoverage(x, u = 1)
}
\arguments{
\item{x}{a \link{GRanges} object.}

\item{u}{numeric, base unit value (e.g 1 => bp, 1E3 => kb, 1E6 => Mb).
By default u = 1 (bp).

\code{GenomicCoverage} returns the total coverage of x.}
}
\description{
Total coverage of genomic intervals
}
\keyword{internal}
