% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{getPrincipalComponent}
\alias{getPrincipalComponent}
\title{Parameters of a principal component}
\usage{
getPrincipalComponent(M, invariant = NULL, pc = "PC1")
}
\arguments{
\item{M}{matrix of read counts (columns = samples, rows = intervals)}

\item{pc}{}
}
\value{
list

getPrincipalComponent returns a \code{list} with the following elements:
\item{SLOPES}{matrix of b values according to the linear model y = b x + a}
\item{INTRCP}{matrix of a values according to the linear model y = b x + a}
}
\description{
Parameters of a principal component
}

