% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_data.R
\docType{data}
\name{hg38_blacklist}
\alias{hg38_blacklist}
\title{ENCODE blacklist - Homo sapiens - GRCh38 / hg38}
\format{An object of class \code{GRanges} of length 1501.}
\usage{
hg38_blacklist
}
\description{
ENCODE blacklist - Homo sapiens - GRCh38 / hg38
}
\keyword{internal}
