% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCSC.R
\name{ImportCpGIslandExt}
\alias{ImportCpGIslandExt}
\title{Import UCSC cpgIslandExt.txt file format}
\usage{
ImportCpGIslandExt(fpath, ...)
}
\arguments{
\item{fpath}{file path or url.}

\item{...}{optional parameters forwarded to the \link{ImportGenomicRanges} function.}
}
\value{
\code{ImportCpGIslandExt} returns a \link{GRanges} object.
}
\description{
Import UCSC cpgIslandExt.txt file format
}
\seealso{
\link{ImportGenomicRanges}
}
\keyword{internal}
