% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicRanges.R
\name{ImportGenomicRanges}
\alias{ImportGenomicRanges}
\title{Import tab delimited text files representing genomic interval data}
\usage{
ImportGenomicRanges(fpath, chr = 1, start = 2, end = 3, strand = NULL,
  seqinfo = NULL, xidx = NULL, xlbl = NULL, header = F, ...)
}
\arguments{
\item{fpath}{path to a text file defining genomic ranges.}

\item{chr}{column for chromsome identifiers (default = 1).}

\item{start}{column for start position (default = 2).}

\item{end}{column for end position (default = 3).}

\item{strand}{column for strand (default = NULL, e.g. none).}

\item{seqinfo}{\link{Seqinfo} object for the considered organism.}

\item{xidx}{list of extra column indexes to be imported.
By default all available columns are imported.}

\item{xlbl}{character vector defining the name of imported extra columns
(default = NULL).}

\item{header}{logical indicating if column names are present (default = F).}

\item{...}{optional parameters forwarded to the \link{read.delim} function.}
}
\value{
\code{ImportGenomicRanges} returns a \link{GRanges} object.
}
\description{
Import tab delimited text files representing genomic interval data
}
\seealso{
\link{ImportCpGIslandExt},
  \link{BuildGeneFeatures},
  \link{GenomicTiling},
  \link{ReadCountMatrix}
}
