% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCSC.R
\name{ImportLiftOverChain}
\alias{ImportLiftOverChain}
\title{Download and import UCSC's liftOver chain files}
\usage{
ImportLiftOverChain(url, path = "")
}
\arguments{
\item{url}{character.}

\item{path}{character.}
}
\value{
\code{ImportLiftOverChain} returns a \link{`Chain-class`} object.
}
\description{
Download and import UCSC's liftOver chain files
}
\keyword{internal}
