% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{JoinColumns}
\alias{JoinColumns}
\title{Merge count data}
\usage{
JoinColumns(x, y)
}
\arguments{
\item{x}{list of read count matrices
(rows = observations, columns = samples or conditions).}

\item{y}{list of read count matrices.}
}
\value{
JoinColumns returns a \code{list}
}
\description{
Merge count data
}
\seealso{
\link{ExtractColumns},
  \link{makeReadCountMatrix},
  \link{makeReadCountsInROI}
}
\keyword{internal}
