% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_stats.R
\name{knn_density}
\alias{knn_density}
\title{knn_density}
\usage{
knn_density(x, k, i = NULL, d = NULL, smoothing = T)
}
\arguments{
\item{x}{numeric matrix representing multivariate observations.}

\item{k}{number of nearest neighbors (i.e. smoothing factor).}

\item{i}{precomputed matrix of nearest neighbor indices.}

\item{d}{precomputed matrix of nearest neighbor distances.}

\item{smoothing}{perfom a knn average smoothing of the density (default = T, recommended).}
}
\value{
knn_density returns a numeric vector.
}
\description{
k-nearest neighbors estimator of density \eqn{P(x) ~ k \over N V}
}
\seealso{
\link{knn_musigma2}
}
