% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_stats.R
\name{knn_mean}
\alias{knn_mean}
\title{Local mean}
\usage{
knn_mean(x, k, i = NULL, d = NULL, smoothing = F)
}
\arguments{
\item{x}{numeric matrix representing multivariate data where rows = observations
and columns = measurement conditions.}

\item{k}{number of nearest neighbors (i.e. smoothing factor).}

\item{i}{precomputed matrix of nearest neighbor indices (optional).}

\item{d}{precomputed matrix of nearest neighbor distances (optional).}
}
\value{
knn_mean returns a \code{matrix} of knn centroids.
}
\description{
compute k-nearest neighbor (knn) centroids (i.e. mean vectors).
}
\seealso{
\link{knn_musigma2},
  \link{knn_density}
}
