% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_stats.R
\name{knn_sd}
\alias{knn_sd}
\title{k-nearest neighbors standard deviation}
\usage{
knn_sd(x, nn.idx, nn.dist, k)
}
\arguments{
\item{x}{numeric matrix representing multivariate observations}

\item{nn.idx}{matrix of nearest neighbor indexes}

\item{nn.dist}{matrix of nearest neighbor distances}

\item{k}{number of nearest neighbors}
}
\value{
knn_sd returns a vector of standard deviations
}
\description{
k-nearest neighbors standard deviation
}
