% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_stats.R
\name{knn_smoothing}
\alias{knn_smoothing}
\title{Apply smoothing function to local values}
\usage{
knn_smoothing(v, i, f = mean)
}
\arguments{
\item{v}{numeric vector.}

\item{i}{precomputed matrix of nearest neighbor indices.}

\item{f}{smoothing function (default = mean).}
}
\value{
\code{knn_smoothing} returns a numeric vector.
}
\description{
Apply smoothing function to local values
}
\seealso{
\link{knn_density},
\link{knn_musigma2}
}
\keyword{internal}
