% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_stats.R
\name{knn_stat}
\alias{knn_stat}
\title{k-nearest neighbor statistics}
\usage{
knn_stat(qry, k, sbj = NULL, FUN = knn_density)
}
\arguments{
\item{qry}{numeric matrix representing multivariate observations}

\item{k}{number of nearest neighbors}

\item{sbj}{}

\item{FUN}{statistical function (default = \link{knn_density})}
}
\value{
knn_stat returns a numeric vector
}
\description{
k-nearest neighbor statistics
}
