% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_stats.R
\name{knn_stats}
\alias{knn_stats}
\title{knn statistics}
\usage{
knn_stats(data, k, smoothing = T)
}
\arguments{
\item{data}{numeric matrix representing multivariate observations}

\item{k}{number of nearest neighbors which corresponds to the smoothing parameter
of estimated statistics (larger values = smoother estimations).}

\item{smoothing}{default = T, recommended}
}
\value{
cdadadr returns a \code{list} with the following elements:
\item{density}{knn density}
\item{variance}{knn variance}
}
\description{
knn statistics
}
