% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_stats.R
\name{knn_values}
\alias{knn_values}
\title{Extract local values from knn index matrix}
\usage{
knn_values(v, i)
}
\arguments{
\item{v}{numeric vector.}

\item{i}{precomputed matrix of nearest neighbor indices.}
}
\value{
knn_values returns a matrix.
}
\description{
Extract local values from knn index matrix
}
\keyword{internal}
