% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{lc2ma}
\alias{lc2ma}
\title{Converts log2(counts) to AM values}
\usage{
lc2ma(x, y = NULL)
}
\arguments{
\item{\code{y}}{}
}
\value{
\code{list} with a and m values
}
\description{
a = (y + x)/2, m = (y - x)
}

