% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_matrixes.R
\name{LoadMappedReads}
\alias{LoadMappedReads}
\title{load/verify a list of mapped read files}
\usage{
LoadMappedReads(aln, ...)
}
\arguments{
\item{aln}{a list with named elements, containing exclusively bam file paths or
\link{GAlignments} objects.}

\item{...}{additional arguments passed to the \link{readGAlignments} function.}
}
\value{
\code{LoadMappedReads} returns a list of \link{GAlignments} objects.
}
\description{
load/verify a list of mapped read files
}
\seealso{
\link{AppendReadCounts},
\link{UpdateReadCounts},
\link{ReadCountMatrix},
\link{MakeReadCounts}
}
\keyword{internal}
