% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RData.R
\name{LoadObj}
\alias{LoadObj}
\title{Load an R object from RData with semi-automated path and file name}
\usage{
LoadObj(obj, path = NULL, name = NULL, need = F, pos = .GlobalEnv,
  overload = F, ...)
}
\arguments{
\item{obj}{an R object previously saved using \link{SaveObj}.}

\item{path}{folder location used to load the R object.
When none is specified it is the current working directory.}

\item{name}{a file name. When omitted (default) this name is automatically set to the
name of the R object being passed as argument.}

\item{need}{logical, when \code{TRUE} LoadObj raises an error if object loading fails
(default = F).}

\item{pos}{the target \link{environment} where the object should be loaded
(default = .GlobalEnv).
See \link{assign} for documentation on the different ways to specify
environments with the \code{pos} paramter.}

\item{overload}{logical. When the R object to be loaded is already defined in the target
environment LoadObj can avoid (overload = F) or force (overload = T) the
reloading of this object.}

\item{...}{optional parameters passed to the \link{readRDS} function.}
}
\value{
logical, \code{TRUE} when object was loaded successfully,
\code{FALSE} otherwise.
}
\description{
Load an R object from RData with semi-automated path and file name
}
\examples{
# R objects are automatically saved in the current working directory
a <- rep("a", 10)
SaveObj(a)
rm(a)
LoadObj(a)
print(a)
file.remove("a.rdata") # delete the file generated for this example

# R objects can be saved and loaded from any path
b <- rep("b", 10)
SaveObj(b, path = "MyRData") # path is automatically created if necessary
b <- NULL
LoadObj(b, path = "MyRData")

# Not yet defined R objects may be created only when no saved version exists
if(! LoadObj(a, "MyRData")) {
  a <- 0
  SaveObj(a, "MyRData")
}
print(a)

# Saved object can be updated by overwriting
a <- 1
SaveObj(a, path = "MyRData")
rm(a)
LoadObj(a)
print(a)
}
\seealso{
\link{SaveObj},
\link{AvailableObj},
\link{readRDS},
\link{reassign}
}
