% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{logCounts}
\alias{logCounts}
\title{Compute log2 transformed counts with different methods}
\usage{
logCounts(cnt, lib.size = 1e+06, chip.con = 1, feat.len = 1000,
  method = "raw", pc = 0)
}
\arguments{
\item{\code{cnt}}{matrix of read counts (columns = samples, rows = intervals)}

\item{\code{chip.con}}{ChIP DNA concentration, default = 1}

\item{\code{method}}{either "raw" (default), "cpm", "rpkm", "voom"}

\item{\code{pc}}{pseudo count, default = 0}
}
\value{
\code{matrix} of log-transformed counts
}
\description{
Compute log2 transformed counts with different methods
}

