% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invariability_DEPRECATED.R
\name{lowVariability}
\alias{lowVariability}
\title{lowVariability}
\usage{
lowVariability(cnt, bg_cols, fit_idx, viz_idx, mincv = 0.05, n = 3,
  k = 64, dithering = 3, max_iter = 100)
}
\arguments{
\item{cnt}{matrix of read counts (rows = observations, columns = samples or conditions).}

\item{bg_cols}{columns with background/control profiles (Input, IgG, etc.)}

\item{fit_idx}{columns considered for low variability analysis}

\item{viz_idx}{columns for 2D plots}

\item{mincv}{minimal component variance}

\item{n}{number of clusters}

\item{k}{number of nearest neighbors}

\item{dithering}{number of dithering passes}

\item{max_iter}{maximum number of iterations}
}
\description{
lowVariability
}
\keyword{internal}
