% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{ma2lc}
\alias{ma2lc}
\title{Converts AM values to log2(counts)}
\usage{
ma2lc(a, m = NULL)
}
\arguments{
\item{\code{m}}{}
}
\value{
\code{list} with x and y values
}
\description{
x = a + m/2, y = a - m/2
}

