% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{make_disk_unif_2D}
\alias{make_disk_unif_2D}
\title{make_disk_unif_2D}
\usage{
make_disk_unif_2D(n, normalized = T)
}
\arguments{
\item{n}{number}

\item{normalized}{logical (default = T)}
}
\value{
matrix
}
\description{
make_disk_unif_2D
}
\keyword{internal}
