% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RData.R
\name{MakePath}
\alias{MakePath}
\title{Concatenate several strings to form a filesystem path}
\usage{
MakePath(..., ext = "")
}
\arguments{
\item{...}{character strings forming a file path}

\item{ext}{file name extension (default = none)}
}
\value{
\code{MakePath} returns a \code{character} value
}
\description{
Concatenate several strings to form a filesystem path
}
\keyword{internal}
