% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{makeReadCountMatrix}
\alias{makeReadCountMatrix}
\title{Count reads overlapping with genomic intervals}
\usage{
makeReadCountMatrix(aln, grg, ...)
}
\arguments{
\item{aln}{list of \link{GAlignments} objects with mapped reads from different samples}

\item{grg}{\link{GRanges} object defining the considered genomic intervals}
}
\value{
makeReadCountMatrix matrix of read counts
(rows = observations, columns = samples or conditions).
}
\description{
Count reads overlapping with genomic intervals
}
\seealso{
\link{DitherCounts},
  \link{JoinColumns},
  \link{ExtractColumns},
  \link{makeReadCountsInROI}
}
