% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{makeReadCountsInROI}
\alias{makeReadCountsInROI}
\title{makeReadCountsInROI}
\usage{
makeReadCountsInROI(ALN, ROI, CNT = NULL, labels = NULL,
  ignore.strand = T, read.extension = 0, ...)
}
\arguments{
\item{ALN}{mapped reads}

\item{ROI}{genomic intervals}

\item{CNT}{previsouly computed read counts}

\item{labels}{sample names}

\item{ignore.strand}{default = T}

\item{read.extension}{not yet implemented}

\item{...}{}
}
\value{
makeReadCountsInROI returns a \code{list} of read count matrixes
}
\description{
makeReadCountsInROI
}
\seealso{
\link{JoinColumns},
  \link{ExtractColumns},
  \link{makeReadCountMatrix}
}
\keyword{internal}
