% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{makeScalingFactors}
\alias{makeScalingFactors}
\title{Scaling factors between read counts in different samples}
\usage{
makeScalingFactors(aln, ctr, bgr = NULL, enr = NULL, use.bgr = NULL,
  use.enr = NULL, replicates = 0, lbls = NULL, dither = T, plot = F,
  ...)
}
\arguments{
\item{aln}{list of \link{GAlignments} objects}

\item{ctr}{\link{GRanges} defining control regions}

\item{bgr}{\link{GRanges} defining background regions (default = all control regions)}

\item{enr}{\link{GRanges} defining enriched regions to be avoided (default = none)}

\item{use.bgr}{logical vector defining the samples from which background regions are taken into account (default = all)}

\item{use.enr}{logical vector defining the samples from which enriched regions are taken into account (default = all)}

\item{replicates}{logical matrix indicating replicate samples (default = none)}

\item{lbls}{character vector of samples names (default = none)}

\item{dither}{logical, activates count dithering (default = T, recommended)}

\item{plot}{logical, activates the drawing of control plots (default = F)}

\item{...}{optional parameters passed to the \link{scalingFactors} function.
These include:
\itemize{

  \item center
  function (see \link{scalingFactors})

  \item pairwise.density
  logical controling the use of pairwise or global count density estimations
  and invariable count selections

  \item d
  density threshold

  \item k
  number of nearest neighbors in kNN density estimation
}
as well as other parameters controlling the optional scatterplots.}
}
\value{
makeScalingFactors returns a \code{list} with the following elements:
\item{counts}{
  matrix of read counts in the control regions
  (columns = samples, rows = genomic intervals)
}
\item{scaling}{
  vector of normalization factor for read counts in each sample
  (\code{counts} columns)
}
\item{pairwise}{
  matrix of pairwise normalization factors between samples
  (\code{counts} columns)
}
\item{low.variation}{
  logical vector indicating the subset of rows assumed invariable by the
  normalization
}
}
\description{
Compute normalization factors based on control regions (for instance genes),
possibly using pre-defined background or enriched regions to further control
the set of genomic intervals that are assumed to exhibit minimal variations.
}

