% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{MatchLogSpace}
\alias{MatchLogSpace}
\title{MatchLogSpace}
\usage{
MatchLogSpace(p, x, b = 2)
}
\arguments{
\item{p}{numeric vector}

\item{x}{sorted vector of unique and strictly positive values}

\item{b}{base of the logarithm (default = 2)}
}
\value{
MatchLogSpace returns a numeric vector
}
\description{
MatchLogSpace
}
\keyword{internal}
