% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{md2xy}
\alias{md2xy}
\title{Convert (a = mean, m = difference) to (x, y) coordinates}
\usage{
md2xy(a, m = NULL)
}
\arguments{
\item{a}{numeric vector, or a matrix with 2 columns or a list containing two vectors.}

\item{m}{numeric vector}
}
\value{
\code{md2xy} returns a list or matrix with (x, y) values.
}
\description{
compute \eqn{x = a + m/2}, and \eqn{y = a - m/2}.
The inverse transformation of \code{md2xy} is \link{xy2md}.
}
\seealso{
\link{xy2md}
}
\keyword{internal}
