% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{MergeCounts}
\alias{MergeCounts}
\title{MergeCounts}
\usage{
MergeCounts(x, y)
}
\arguments{
\item{x}{list of read count matrices
(rows = observations, columns = samples or conditions).}

\item{y}{list of read count matrices.}
}
\value{
DitherCounts returns a \code{list}
}
\description{
MergeCounts
}
