% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mkdir}
\alias{mkdir}
\title{Make a destination path creating folders when necessary}
\usage{
mkdir(x)
}
\arguments{
\item{x}{character value indicating a file system path.}
}
\value{
see \link{system}.
}
\description{
Make a destination path creating folders when necessary
}
\keyword{internal}
