% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_data.R
\docType{data}
\name{mm10_blacklist}
\alias{mm10_blacklist}
\title{ENCODE blacklist - Mus musculus - GRCm38 / mm10}
\format{An object of class \code{GRanges} of length 3155.}
\usage{
mm10_blacklist
}
\description{
ENCODE blacklist - Mus musculus - GRCm38 / mm10
}
\keyword{internal}
