% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_gmm.R
\name{mv_emloop}
\alias{mv_emloop}
\title{Multivariate expectation-maximization}
\usage{
mv_emloop(x, theta, p_fun = mv_pdf, mle_fun = mv_mle, max_iter = 100)
}
\arguments{
\item{x}{numeric matrix representing multivariate data where rows = observations
and columns = samples or conditions.}

\item{theta}{list of multivariate distribution parameters}

\item{p_fun}{probability density function (default = \link{mv_pdf})}

\item{mle_fun}{maximum likelyhood estimator function (default = \link{mv_mle})}

\item{max_iter}{maximum number of iterations (default = 100)}
}
\value{
\code{mv_emloop} returns a list with the following elements:
\item{theta}{list of multivariate distribution parameters}
\item{p}{vector of probabilities}
}
\description{
Multivariate expectation-maximization
}
\seealso{
\link{mv_maximization},
\link{mv_expectation},
\link{mv_gmm}
}
\keyword{internal}
