% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvfit.R
\name{mv_expectation}
\alias{mv_expectation}
\title{Multivariate expectation}
\usage{
mv_expectation(x, theta, p_fun = mv_pdf)
}
\arguments{
\item{x}{numeric matrix representing multivariate observations}

\item{theta}{list of multivariate distribution parameters}

\item{p_fun}{probability density function (default = \link{mv_pdf})}
}
\value{
mv_expectation returns a vector of probabilities
}
\description{
Multivariate expectation
}
