% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvfit.R
\name{mv_maximization}
\alias{mv_maximization}
\title{Multivariate maximization}
\usage{
mv_maximization(x, theta, p, mle_fun = mv_mle)
}
\arguments{
\item{x}{numeric matrix representing multivariate observations}

\item{theta}{list of multivariate distribution parameters}

\item{p}{vector of probabilities}

\item{mle_fun}{maximum likelyhood estimator function (default = \link{mv_mle})}
}
\value{
mv_maximization returns a list of multivariate distribution parameters
}
\description{
Multivariate maximization
}
