% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvfit.R
\name{mv_pdf}
\alias{mv_pdf}
\title{Multivariate normal distribution: probability density function}
\usage{
mv_pdf(x, theta)
}
\arguments{
\item{x}{numeric matrix representing multivariate observations}

\item{theta}{list of normal distribution parameters}
}
\value{
mv_pdf returns a vector of probabilities
}
\description{
Multivariate normal distribution: probability density function
}
