% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{NonZeroCounts}
\alias{NonZeroCounts}
\title{remove observations with missing values}
\usage{
NonZeroCounts(cnt)
}
\arguments{
\item{cnt}{matrix of read counts
(rows = observations, columns = measurement conditions).}
}
\value{
\code{NonZeroCounts} returns a matrix.
}
\description{
remove observations with missing values
}
\seealso{
\link{FiniteValues},
\link{DetectCounts},
\link{MakeReadCountMatrix}
}
