% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalizeByCenterOfDifferences}
\alias{normalizeByCenterOfDifferences}
\title{Normalization based on center of differences}
\usage{
normalizeByCenterOfDifferences(M, invariant = NULL, method = c("mode",
  "median", "mean"), cod.matrix = NULL)
}
\arguments{
\item{M}{matrix of read counts (columns = samples, rows = intervals)}

\item{invariant}{logical}

\item{method}{either "mode" (default), "median" or "mean"}
}
\value{
normalizeByCenterOfDifferences returns a \code{list} with the following elements:
\item{X}{matrix of normalized read counts}
\item{f}{matrix of pairwise normalization factors}
}
\description{
Normalization based on center of differences
}
\note{
This is a deprecated function that will be removed in future versions
}

