% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalizeByPrincipalComponent}
\alias{normalizeByPrincipalComponent}
\title{PCA-based normalization}
\usage{
normalizeByPrincipalComponent(M, pc)
}
\arguments{
\item{M}{matrix of read counts (columns = samples, rows = intervals)}

\item{pc}{}
}
\value{
matrix
}
\description{
PCA-based normalization
}

