% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRD_V3.R
\name{NormalizeCountMatrix}
\alias{NormalizeCountMatrix}
\title{Apply BRD scaling factors}
\usage{
NormalizeCountMatrix(x, brd, as.log2 = FALSE)
}
\arguments{
\item{x}{matrix of read counts (rows = observations, columns = samples or conditions).}

\item{brd}{result of a prior call to the \link{BRD} function.}

\item{as.log2}{logical (default = FALSE, no).}
}
\value{
\code{NormalizeCountMatrix} returns a \code{matrix}.
}
\description{
Apply precomputed BRD normalization factors to a read count matrix.
Use \code{as.log2 = TRUE} if you want to provide a matrix of log2
transformed read counts instead of raw read counts.
}
\seealso{
\link{BRD},
  \link{ScalingFactors},
  \link{ReadCountMatrix}
}
