% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_data.R
\docType{data}
\name{NSC_K27M_COUNTS}
\alias{NSC_K27M_COUNTS}
\alias{NSC_K27M_METADATA}
\alias{NSC_K27M_SPIKEIN}
\title{H3K27me3 ChIP-seq in mouse Neuronal Stem Cell (NSCs) - READ COUNTS}
\format{An object of class \code{list} of length 6.}
\usage{
NSC_K27M_COUNTS

NSC_K27M_METADATA

NSC_K27M_SPIKEIN
}
\description{
NSC_K27M_COUNTS: Precomputed read count matrixes (mm10)

NSC_K27M_METADATA: Annotation of experimental conditions

NSC_K27M_SPIKEIN: Number of Drosophila reads (spike-in)

H3K27me3 ChIP-seq in mouse Neuronal Stem Cell (NSCs) - METADATA

H3K27me3 ChIP-seq in mouse Neuronal Stem Cell (NSCs) - SPIKEIN
}
\details{
This ChIP-seq dataset was generated by reproducing experiments presented
in our study about the role of the H3.3K27M histone mutation in DIPG
(Mohammad et al., 2017), but adding spike-in with Drosophila chromatin
in the ChIP-seq protocol.
Briefly, H3K27me3 ChIP-seq experiments were performed in mouse NSCs, in
presence or absence of expression of the H3.3K27M histone mutant (K27M)
as well as with or without treatment with EPZ6438 (EPZ), an inhibitor of the
catalytic reaction producing the H3K27me3 histone mark.
}
\references{
Mohammad F., Weissmann S., Leblanc B., Pandey D.P. et al.
EZH2 is a potential therapeutic target for H3K27M-mutant pediatric gliomas.
Nature Medicine 2017 \url{http://dx.doi.org/10.1038/nm.4293}
}
\keyword{datasets}
\keyword{internal}
