% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_groups_2D}
\alias{plot_groups_2D}
\title{2D scatter plot with groups}
\usage{
plot_groups_2D(x, p = NULL, posterior = 0, contrast = 1, clr = NULL,
  alpha = 0.1, ...)
}
\arguments{
\item{x}{numeric matrix, the two first columns being used as x and y coordinates}

\item{p}{matrix of group membership probabilities}

\item{posterior}{membership probability threshold}

\item{contrast}{membership probability contrast factor}

\item{clr}{vector of colors}

\item{alpha}{color transparency}

\item{...}{}
}
\value{
plot_groups_2D returns a vector of colors
}
\description{
2D scatter plot with groups
}
