% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_samples}
\alias{plot_samples}
\title{plot_samples}
\usage{
plot_samples(x, idx = 1:2, symetric = T, cex = 0.5, ...)
}
\arguments{
\item{x}{matrix}

\item{idx}{vector}

\item{symetric}{logical}

\item{cex}{numeric}

\item{...}{}
}
\description{
plot_samples
}
\keyword{internal}
