% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_sources_2D}
\alias{plot_sources_2D}
\title{plot 2D normal distributions}
\usage{
plot_sources_2D(theta, components = NULL, p = 0.5, clr = NULL,
  center = F, ...)
}
\arguments{
\item{theta}{list of multivariate distribution parameters}

\item{components}{components to be represented (default = 1:2)}

\item{p}{probability boundary of represented distributions}

\item{clr}{vector of colors}

\item{center}{logical (default = F)}

\item{...}{}
}
\description{
plot 2D normal distributions
}
