% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBRD.R
\name{PlotBRD}
\alias{PlotBRD}
\title{PlotBRD}
\usage{
PlotBRD(brd, with.axes = T, with.legend = T, res = 200)
}
\arguments{
\item{brd}{result from \link{BRD}}

\item{with.axes}{logical (default = T)}

\item{with.legend}{logical (default = T)}

\item{res}{integer}
}
\description{
PlotBRD
}
\keyword{internal}
