% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBRD_V3.R
\name{PlotCountDistributions}
\alias{PlotCountDistributions}
\title{Plot empirical distributions of read counts}
\usage{
PlotCountDistributions(x, ...)
}
\arguments{
\item{x}{a read count matrix, preferably dithered and log2 transformed.}

\item{...}{optional arguments forwarded to the \link{SideBySide} function.}
}
\description{
Plot empirical distributions of read counts
}
\examples{
# Simulation of a read count matrix with 3 populations of 10000 observations
p <- c(10000, 10000, 10000)
m <- DefineSimulation(
  chip = 5, patterns = c("^", "v"), enrichment = c(1.0, 3), replicate = 2
)
r <-  MakeSimulation(p = p, m = m)

grp <- r$group # Population memberships
cnt <- r$data  # Simulated counts

# Prepare figure layout and graphic options
layout(matrix(1:4, 2, 2, byrow = TRUE))
par(pch = 20)

# Show the empirical distribution of simulated populations
l2c <- log2(DitherCounts(cnt)) # Dithering and log2 transformation
xyl <- range(l2c[FiniteValues(l2c), ])

r <- PlotCountDistributions(l2c, ylim = xyl, main = "Total")
for(i in sort(unique(grp))) {
  main <- ifelse(i == 1, "Invariable subset", paste("Variable subset", i - 1))
  r <- PlotCountDistributions(l2c[grp == i, ], ylim = xyl, main = main)
}
}
\seealso{
\link{DitherCounts},
  \link{ReadCountMatrix}
}
