% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotHistograms}
\alias{plotHistograms}
\title{plotHistograms}
\usage{
plotHistograms(x, y, bins = 100, xlim = NULL, log = F, rel = F, ...)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{bins}{integer}

\item{xlim}{range}

\item{log}{logical}

\item{rel}{logical}

\item{...}{}
}
\description{
plotHistograms
}
\keyword{internal}
