% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{PlotQuickShift}
\alias{PlotQuickShift}
\title{PlotQuickShift}
\usage{
PlotQuickShift(x, g, new = T, ...)
}
\arguments{
\item{x}{matrix}

\item{g}{graph}

\item{...}{}
}
\description{
PlotQuickShift
}
\keyword{internal}
