% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invariability_prototype.R
\name{populations}
\alias{populations}
\title{partition of multivariate observations}
\usage{
populations(x, tx = NULL, p, ns, max_iter = 100)
}
\arguments{
\item{x}{numeric matrix representing multivariate data where rows = observations
and columns = samples or conditions.}

\item{tx}{numeric matrix representing transformed observations (default = x).}

\item{p}{statistical score associated to observations (e.g. estimated density).}

\item{ns}{number of clusters.}

\item{max_iter}{maximum number of iterations for expectation-maximization}
}
\value{
populations returns a \code{list} with the following elements:
\item{grp}{list of clusters resulting from function \link{QuickShiftCutClusters}}
\item{qstree}{QuickShift tree}
\item{theta}{list of multivariate distribution parameters}
}
\description{
partition of multivariate observations
}
\keyword{internal}
