% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickshift.R
\name{quickShift}
\alias{quickShift}
\title{quickshift algorithm (Vedaldi & Soatto, 2008)}
\usage{
quickShift(x, d, plot = F)
}
\arguments{
\item{x}{numeric matrix representing multivariate observations}

\item{d}{statistical score associated to observations (e.g. estimated density)}

\item{plot}{logical (default = F)}
}
\value{
quickShift returns a graph object (igraph package)
}
\description{
quickshift algorithm (Vedaldi & Soatto, 2008)
}
