% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickshift.R
\name{QuickShiftClustering}
\alias{QuickShiftClustering}
\title{QuickShiftClustering}
\usage{
QuickShiftClustering(x, d, n, ...)
}
\arguments{
\item{x}{numeric matrix representing multivariate data where rows = observations
and columns = samples or conditions.}

\item{d}{numeric vector representing the density estimation at each observation.}

\item{n}{number of clusters}

\item{...}{}
}
\value{
QuickShiftClustering returns a \code{list} with the following elements
}
\description{
QuickShiftClustering
}
\seealso{
\link{BRD},
  \link{QuickShift},
  \link{QuickShiftCutClusters}
}
