% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickshift.R
\name{QuickShiftSearchClusters}
\alias{QuickShiftSearchClusters}
\title{QuickShift root and top level branches}
\usage{
QuickShiftSearchClusters(i = 0, d = 0, dmax)
}
\arguments{
\item{i}{node index}

\item{d}{depth level}

\item{dmax}{maximum depth}
}
\value{
QuickShiftSearchClusters returns a \code{data.frame}
}
\description{
QuickShift root and top level branches
}
