% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{Random2D}
\alias{Random2D}
\title{Random2D}
\usage{
Random2D(n, p, v = NULL)
}
\arguments{
\item{n}{number}

\item{p}{probabilities}

\item{v}{values}
}
\value{
matrix
}
\description{
Random2D
}
\keyword{internal}
