% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadCountSimulation.R
\name{RandomCounts}
\alias{RandomCounts}
\title{Random counts}
\usage{
RandomCounts(n, d, k = NULL, p = NULL, v = NULL, extended = FALSE)
}
\arguments{
\item{n}{number of observations (rows).}

\item{d}{number of dimensions (columns).}

\item{k}{by default count values are sampled in [1, \code{k}].
This argument is ignored when arguments \code{p} or \code{v} are provided.}

\item{p}{vector of count probabilities (default = uniform).}

\item{v}{vector of count values (default = \code{1:k})}

\item{extended}{logical (default = FALSE)}
}
\value{
\code{RandomCounts} returns a list.
}
\description{
Generate pseudo random count values
}
\keyword{internal}
