% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rankstat}
\alias{rankstat}
\title{rank statistics}
\usage{
rankstat(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
\code{rankstat} returns the rank statistics of x.
}
\description{
compute the following rank statistics
\deqn{(rank(x) - 0.5) / N} where N = length(x).
}
\keyword{internal}
