% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RData.R
\name{rdata2tsv}
\alias{rdata2tsv}
\title{Convert RData file to tab delimited file}
\usage{
rdata2tsv(path, name, col.names = T, row.names = F)
}
\arguments{
\item{path}{character}

\item{name}{character}

\item{col.names}{logical}

\item{row.names}{logical}
}
\description{
Convert RData file to tab delimited file
}
\keyword{internal}
