% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadCountMatrix.R
\name{ReadCountMatrix}
\alias{ReadCountMatrix}
\title{count reads overlapping with genomic intervals}
\usage{
ReadCountMatrix(lst, grg, cnt = NULL, ...)
}
\arguments{
\item{lst}{list of \link{GAlignments} objects with mapped reads from different samples}

\item{grg}{\link{GRanges} object defining the considered genomic intervals}
}
\value{
\code{ReadCountMatrix} returns a matrix of read counts with
rows = observations and columns = measurement conditions.
}
\description{
count reads overlapping with genomic intervals
}
\seealso{
\link{NonZeroCounts},
\link{DitherCounts},
\link{JoinColumns},
\link{ExtractColumns},
\link{makeReadCountsInROI}
}
