% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RData.R
\name{reassign}
\alias{reassign}
\title{Reassign object to a different environment}
\usage{
reassign(obj, pos, src = .GlobalEnv, keep = F)
}
\arguments{
\item{obj}{an R object}

\item{pos}{the target \link{environment} where the object should be reassigned.
See \link{assign} for documentation on the different ways to specify
environments with the \code{pos} paramter.}

\item{src}{the source \link{environment} where the object is currently located
(default = .GlobalEnv).
See \link{assign} for documentation on the different ways to specify
environments with the \code{pos} paramter.}

\item{keep}{logical indicating if the original object should be removed or preserved
at its current location (defautl = F, remove).}
}
\description{
Reassign object to a different environment
}
\seealso{
\link{LoadObj},
\link{SaveObj},
\link{AvailableObj},
\link{assign},
\link{environment}
}
\keyword{internal}
