% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_matrixes.R
\name{RemoveColumns}
\alias{RemoveColumns}
\title{Remove count data}
\usage{
RemoveColumns(x, lst)
}
\arguments{
\item{x}{list of read count matrices
(rows = observations, columns = measurement conditions).}

\item{lst}{names or indices of columns to be removed.}
}
\value{
\code{RemoveColumns} returns a list of read count matrices.
}
\description{
Remove count data
}
\seealso{
\link{JoinColumns},
\link{ExtractColumns},
\link{MakeReadCounts}
}
\keyword{internal}
