% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tightrope.R
\name{RemoveOptions}
\alias{RemoveOptions}
\title{** RESERVED FOR INTERNAL USE **}
\usage{
RemoveOptions()
}
\description{
Remove global options of the Tightrope package from the R environment
}
\keyword{internal}
