% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{RotationMatrix2D}
\alias{RotationMatrix2D}
\title{RotationMatrix2D}
\usage{
RotationMatrix2D(alpha)
}
\arguments{
\item{alpha}{angle}
}
\value{
matrix
}
\description{
RotationMatrix2D
}
\keyword{internal}
