% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RData.R
\name{SaveObj}
\alias{SaveObj}
\title{Save an R object as RData with semi-automated path and file name}
\usage{
SaveObj(obj, path = NULL, name = NULL, ...)
}
\arguments{
\item{obj}{an R object to be saved as RData.}

\item{path}{folder location used to save the R object.
When none is specified it is the current working directory.}

\item{name}{a file name. When omitted (default) this name is automatically set to the
name of the R object being passed as argument.}

\item{...}{optional parameters passed to the \link{saveRDS} function.}
}
\description{
Save an R object as RData with semi-automated path and file name
}
\examples{
# R objects are automatically saved in the current working directory
a <- rep("a", 10)
SaveObj(a)
rm(a)
LoadObj(a)
print(a)
file.remove("a.rdata") # delete the file generated for this example

# R objects can be saved and loaded from any path
b <- rep("b", 10)
SaveObj(b, path = "MyRData") # path is automatically created if necessary
b <- NULL
LoadObj(b, path = "MyRData")

# Not yet defined R objects may be created only when no saved version exists
if(! LoadObj(a, "MyRData")) {
  a <- 0
  SaveObj(a, "MyRData")
}
print(a)

# Saved object can be updated by overwriting
a <- 1
SaveObj(a, path = "MyRData")
rm(a)
LoadObj(a)
print(a)
}
\seealso{
\link{LoadObj},
\link{AvailableObj},
\link{saveRDS},
\link{save.image},
\link{reassign}
}
