% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{scPlotCounts}
\alias{scPlotCounts}
\title{Scatterplot with points colored according to kNN density estimation}
\usage{
scPlotCounts(X, col = 0, grp = NULL, grp.col = 0, zline = F,
  xyline = F, k = 256, ...)
}
\arguments{
\item{\code{X}}{matrix with two columns of numeric values}

\item{\code{col}}{index of color gradient (0 = grey, 1 = red, 2 = green, 3 = blue)}

\item{\code{grp}}{logical vector defining an highlighted subset of points}

\item{\code{grp.col}}{index of color gradient used for the grp subset}

\item{\code{zline}}{logical, activates the drawing of line y = 0 when TRUE}

\item{\code{xyline}}{logical, activates the drawing of line y = x when TRUE}

\item{\code{k}}{number of nearest neighbors for kNN density estimation}

\item{\code{...}}{optional parameters passed to the plot function}
}
\value{
produce a scatter plot in the active graphic device
}
\description{
Scatterplot with points colored according to kNN density estimation
}

