% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{SimulatedCounts}
\alias{SimulatedCounts}
\title{Random counts representing 3 populations}
\usage{
SimulatedCounts(n, d = 3, mu = 5, sigma = 3, poorcounts = list(),
  invariable = list(), variable = list(), as.counts = T)
}
\arguments{
\item{n}{number of observations}

\item{d}{number of conditions (dimensions)}

\item{mu}{average read counts (provided as log transformed value)}

\item{sigma}{}

\item{poorcounts}{parameters of the barely detectable subpopulation}

\item{invariable}{parameters of the subpopulation generating background or low variability
counts}

\item{variable}{parameters of the subpopulation showing significant variations}

\item{as.counts}{return counts (default = T, yes)}
}
\value{
\code{SimulatedCounts} returns a list.
}
\description{
Random counts representing 3 populations
}
\keyword{internal}
