% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{SkewedProbabilities}
\alias{SkewedProbabilities}
\title{SkewedProbabilities}
\usage{
SkewedProbabilities(k, beta = -1/8, gamma = 0.5, sx = -1, sy = 1,
  plot = F)
}
\arguments{
\item{k}{resolution}

\item{beta}{exponent}

\item{gamma}{mix}

\item{sx}{+1 or -1}

\item{sy}{+1 or -1}

\item{plot}{default = F}
}
\value{
\code{SkewedProbabilities} returns a vector of probabilities.
}
\description{
SkewedProbabilities
}
\keyword{internal}
