% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{SquareGrid}
\alias{SquareGrid}
\title{SquareGrid}
\usage{
SquareGrid(k, d, radius = -1, subset = NULL)
}
\arguments{
\item{k}{integer}

\item{d}{dimension}

\item{radius}{distance of exclusion}

\item{subset}{vector of indices}
}
\value{
list
}
\description{
SquareGrid
}
\keyword{internal}
