% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.test.R
\name{statusByFoldChange}
\alias{statusByFoldChange}
\title{Categorize differences according to fold-changes}
\usage{
statusByFoldChange(a, m, pval, padj, bg.tail, cmp, bg = log2(1.5),
  h = log2(2), amin = 4)
}
\arguments{
\item{amin}{}
}
\value{
\code{data.frame}
}
\description{
Categorize differences according to fold-changes
}

