% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.test.R
\name{statusByPValue}
\alias{statusByPValue}
\title{Categorize differences according to p-values}
\usage{
statusByPValue(a, m, pval, padj, bg.tail, cmp, bg = 0.2, h = 0.01,
  amin = 4)
}
\arguments{
\item{amin}{}
}
\value{
\code{data.frame}
}
\description{
Categorize differences according to p-values
}

