% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_matrixes.R
\name{StopOnError}
\alias{StopOnError}
\title{StopOnError}
\usage{
StopOnError(msg, chk, x)
}
\arguments{
\item{msg}{error message.}

\item{chk}{logical vector.}

\item{x}{character vector.}
}
\value{
\code{StopOnError} stops execution in case of error or returns FALSE.
}
\description{
StopOnError
}
\keyword{internal}
