% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{SX2Y}
\alias{SX2Y}
\title{SX2Y}
\usage{
SX2Y(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\value{
\code{SX2Y} returns x rescaled such that range(x) = range(y).
}
\description{
rescale x values linearly to match the range of y.
}
\seealso{
\link{rankstat},
  \link{S01}
}
\keyword{internal}
