% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tightrope.R
\name{Tightrope}
\alias{Tightrope}
\title{Global options for Tightrope functions}
\usage{
Tightrope(...)
}
\arguments{
\item{...}{Any of the following arguments:}
}
\description{
This function sets the default value of arguments used by the main functions
of the Tightrope package.
}
\seealso{
\link{ScatterPlot},
  \link{SideBySide}
}
