% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{translation_2D}
\alias{translation_2D}
\title{translation_2D}
\usage{
translation_2D(v, dv)
}
\arguments{
\item{v}{matrix}

\item{dv}{vector}
}
\value{
matrix
}
\description{
translation_2D
}
\keyword{internal}
