% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RData.R
\name{tsv2rdata}
\alias{tsv2rdata}
\title{Convert tab delimited file to RData}
\usage{
tsv2rdata(path, name)
}
\arguments{
\item{path}{character}

\item{name}{character}
}
\description{
Convert tab delimited file to RData
}
\keyword{internal}
