% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_arguments.R
\name{VectorArgs}
\alias{VectorArgs}
\title{** RESERVED FOR INTERNAL USE **}
\usage{
VectorArgs(lst, from = NULL, size = NULL)
}
\description{
Standardize the length of vector arguments.
}
\keyword{internal}
