% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_matrixes.R
\name{VerifyInputs}
\alias{VerifyInputs}
\title{VerifyInputs}
\usage{
VerifyInputs(cnt, grg)
}
\arguments{
\item{cnt}{a numeric matrix is expected.}

\item{grg}{a GRanges object is expected.}
}
\value{
\code{VerifyInputs} stops execution in case of invalid input or returns TRUE.
}
\description{
VerifyInputs
}
\keyword{internal}
