% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRD_V3.R
\name{bg_ranking}
\alias{bg_ranking}
\title{Average ranking for control measurements}
\usage{
bg_ranking(x, by, controls)
}
\arguments{
\item{x}{matrix of read counts
(rows = observations, columns = measurement conditions).}

\item{by}{grouping index.}

\item{controls}{columns corresponding to control measurements (Input, IgG, etc.)}
}
\value{
\code{bg_ranking} returns group identifiers ordered by decreasing background
intensity, meaning the average difference between control enrichment
conditions.
}
\description{
Average ranking for control measurements
(e.g. input DNA, ChIP-seq performed with IgG or in KO conditions).
}
\seealso{
\link{BRD}
}
\keyword{internal}
