% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRD_V3.R
\name{BRD}
\alias{BRD}
\title{Background Read Density}
\usage{
BRD(cnt, controls, sampling = NULL, smobs = TRUE, dither = 5,
  zscore = TRUE, bins = 705, smoothing = 25, bdt = c(0.15, 0.05),
  ncl = 1, mincs = 50, progress = FALSE)
}
\arguments{
\item{cnt}{matrix of read counts
(rows = observations, columns = measurement conditions).}

\item{controls}{count columns corresponding to control measurements (Input, IgG, etc.).}

\item{smobs}{subtract a mean count value for each observation (default = TRUE, recommended).}

\item{dither}{number of replicates of the count dithering performed by \link{DitherCounts}:
1 = single, 2 = duplicate, 3 = triplicate, etc. (default = 5).}

\item{zscore}{transform read count projections into z-scores (default = TRUE, recommended).}

\item{bins}{number of bins per principal component for density estimations
(default = 500).}

\item{smoothing}{number of consecutive bins for local average smoothing of estimated densities
(default = 10).}

\item{bdt}{numeric vector of length 2 defining background density thresholds both
expressed as proportions between 0 and 1.
\code{bdt[1]} specifies the global threshold used to discard observations
with low density prior to clustering.
\code{bdt[2]} determines the maximum density loss allowed
when selecting core observations relatively to the local maximum density
in each cluster, and thus defining the background candidates.
By default the value of \code{bdt} is \code{c(0.2, 0.05)} meaning that,
in terms of density percentiles, the bottom 20 percents will be filtered out
before clustering and only the top 5 percents can be selected as background
candidates among each cluster.}

\item{ncl}{number of clusters (density modes) to be distinguished.}

\item{mincs}{minimum size of cluster cores, as number of observations.}
}
\value{
\code{BRD} returns a list with the following elements:
\item{parameters}{call parameters of the function.}
\item{status}{execution status.}
\item{nonzero}{indices of initial observations with count > 0.}
\item{dred}{
  dimensionality-reduced non-zero observations.
}
\item{subsets}{
  partition of non-zero observations into background candidate subsets.
}
\item{populations}{
  summary of the core population in each subset.
}
\item{theta}{
  fitted distribution parameters for each core population.
}
\item{log2counts}{
  dithered and log2 transformed counts.
}
\item{normfactors}{
  BRD normalization factors.
}
}
\description{
The \code{BRD} function searches for background candidates based on the
provided read count matrix and uses these background candidates to estimate
normalization factors.
}
\seealso{
\link{PlotBRD},
  \link{BackgroundCandidates},
  \link{ScalingFactors},
  \link{NormalizeCountMatrix},
  \link{ReadCountMatrix}
}
