% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRD.R
\name{BRD}
\alias{BRD}
\title{Background Read Density}
\usage{
BRD(cnt, controls = NULL, dither = 3, smobs = T, cvt = 0.5, npc = NA,
  zscore = T, knn = 200, bdt = c(0.5, 0.1), ncl = 3, mincs = 300,
  progress = F)
}
\arguments{
\item{cnt}{matrix of read counts
(rows = observations, columns = measurement conditions).}

\item{controls}{columns corresponding to control measurements (Input, IgG, etc.).}

\item{dither}{number of replicates for count dithering:
1 = single (default), 2 = duplicate, 3 = triplicate, etc.}

\item{smobs}{subtract the mean count value of each observation (logical, default = F).}

\item{cvt}{percentage of variance to be ignored.
The value of \code{cvt} determines the stringency of the dimensionality
reduction which is performed by principal component analysis using
\link{prcomp}.
Dimensions retained after projection of the read count dataset on its
principal components will preserve \code{100 * (1 - cvt)} percents
of the initial variance.
For instance, after dimensionality reduction with the default setting
(\code{cvt} = 0.5), approximately 50 percents of the initial variance
will remain in the reduced data.
Setting \code{cvt} = 0 bypasses the dimensionality reduction and performs
density estimations in the original multivariate measurement space
of the dataset.}

\item{npc}{number of dimensions retained after projection on principal components.
When \code{npc} is specified, the \code{cvt} parameter is ignored.}

\item{zscore}{transform counts into z-scores (logical, default = T).}

\item{knn}{number of nearest neighbors, which corresponds to the smoothing parameter
of estimated densities (larger values = smoother).}

\item{bdt}{numeric vector of length 2 defining background density thresholds both
expressed as proportions between 0 and 1.
(\code{bdt[1]}) specifies the global threshold used to discard observations
with low density prior to clustering.
(\code{bdt[2]}) determines the maximum density loss allowed
when selecting core observations relatively to the local maximum density
in each cluster, and thus defining the candidate background populations.
By default the value of \code{bdt} is \code{c(0.5, 0.1)} meaning that,
in terms of density values and relative to the total population of
observations, the bottom 50% will be filtered out before clustering
and only the top 10% can be selected as background candidates within each
cluster.}

\item{ncl}{number of clusters for partitioning observations.}

\item{mincs}{minimum size of each cluster, as number of observations.}

\item{progress}{show progress bar (logical, default = F).}
}
\value{
\code{BRD} returns a \code{list} with the following elements:
\item{parameters}{call parameters of the function.}
\item{status}{execution status.}
\item{nonzero}{indices of initial observations with count > 0.}
\item{dred}{
  dimensionality-reduced non-zero observations
  (result from the \link{CDaDaDR} function).
}
\item{subsets}{
  partition of non-zero observations into background candidate subsets.
}
\item{populations}{
  summary of core populations.
}
\item{theta}{
  fitted distribution parameters for each core candidate populations.
}
\item{log2counts}{
  dithered and log2 transformed counts.
}
}
\description{
Background Read Density
}
\seealso{
\link{CPSES},
  \link{QuickShiftClustering},
  \link{CDaDaDR}
}
