% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRD.R
\name{BRD}
\alias{BRD}
\title{background read density estimation}
\usage{
BRD(cnt, controls, smobs = F, dither = 3, npc = 2, zscore = T,
  knn = 200, rare = 0.01, method = "pca", bdt = c(0.5, 0.05), ncl = 2,
  mincs = 100, progress = F)
}
\arguments{
\item{cnt}{matrix of read counts
(rows = observations, columns = measurement conditions).}

\item{controls}{count columns corresponding to control measurements (Input, IgG, etc.).}

\item{smobs}{subtract the mean count value of each observation (logical, default = F).}

\item{dither}{number of replicates of the count dithering performed by \link{DitherCounts}:
1 = single (default), 2 = duplicate, 3 = triplicate, etc.}

\item{npc}{number of dimensions retained after projection on principal or independent
components.}

\item{zscore}{transform read count projections into z-scores (logical, default = T).}

\item{knn}{number of nearest neighbors, which corresponds to the smoothing parameter
of densities estimated by \link{knn_density} (larger knn values = smoother).}

\item{rare}{mark rare observations corresponding to exceptionally low density values
(default = 0.01).}

\item{method}{projection method, either "pca" (default) performed by \link{prcomp},
or "ica" performed by \link{icafast}.}

\item{bdt}{numeric vector of length 2 defining background density thresholds both
expressed as proportions between 0 and 1.
\code{bdt[1]} specifies the global threshold used to discard observations
with low density prior to clustering.
\code{bdt[2]} determines the maximum density loss allowed
when selecting core observations relatively to the local maximum density
in each cluster, and thus defining the candidate background populations.
By default the value of \code{bdt} is \code{c(0.5, 0.05)} meaning that,
in terms of density percentiles, the bottom 30 percents will be filtered out
before clustering and only the top 10 percents can be selected as background
candidates among each cluster.}

\item{ncl}{number of clusters partitioning the population of observations with density
above \code{bdt[1]} percents.}

\item{mincs}{minimum size of cluster cores, as number of observations.}

\item{progress}{show progress bar (logical, default = F).}
}
\value{
\code{BRD} returns a list with the following elements:
\item{parameters}{call parameters of the function.}
\item{status}{execution status.}
\item{nonzero}{indices of initial observations with count > 0.}
\item{dred}{
  dimensionality-reduced non-zero observations
  (result from the \link{CDaDaDR} function).
}
\item{subsets}{
  partition of non-zero observations into background candidate subsets.
}
\item{populations}{
  summary of the core population in each subset.
}
\item{theta}{
  fitted distribution parameters for each core population.
}
\item{log2counts}{
  dithered and log2 transformed counts.
}
}
\description{
background read density estimation
}
\seealso{
\link{PlotBRD},
\link{CDaDaDR},
\link{QuickShiftClustering},
\link{MakeReadCountMatrix}
}
