% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRD.R
\name{BRD}
\alias{BRD}
\title{Background Read Density}
\usage{
BRD(cnt, controls = NULL, dither = 3, smobs = T, cvt = 0.5, npc = NA,
  zscore = T, knn = 200, cst = 0.5, ncl = 3, mincs = 300,
  progress = F)
}
\arguments{
\item{cnt}{matrix of read counts
(rows = observations, columns = measurement conditions).}

\item{controls}{columns corresponding to control measurements (Input, IgG, etc.).}

\item{dither}{number of replicates for count dithering:
1 = single (default), 2 = duplicate, 3 = triplicate, etc.}

\item{smobs}{subtract the mean count value of each observation (logical, default = F).}

\item{cvt}{percentage of variance to be ignored.
The value of \code{cvt} determines the stringency of the dimensionality
reduction which is performed by principal component analysis using
\link{prcomp}.
Dimensions retained after projection of the read count dataset on its
principal components will preserve \code{100 * (1 - cvt)} percents
of the initial variance.
For instance, after dimensionality reduction with the default setting
(\code{cvt} = 0.5), approximately 50 percents of the initial variance
will remain in the reduced data.
Setting \code{cvt} = 0 bypasses the dimensionality reduction and performs
density estimations in the original multivariate measurement space
of the dataset.}

\item{npc}{number of dimensions retained after projection on principal components.
When \code{npc} is specified, the \code{cvt} parameter is ignored.}

\item{zscore}{transform counts into z-scores (logical, default = T).}

\item{knn}{number of nearest neighbors, which corresponds to the smoothing parameter
of estimated densities (larger values = smoother).}

\item{cst}{core subset density threshold.}

\item{ncl}{number of clusters used to partition the core subset.}

\item{mincs}{minimum size of each cluster, as number of observations.}

\item{progress}{show progress bar (logical, default = F).}
}
\value{
\code{BRD} returns a \code{list} with the following elements:
\item{parameters}{call parameters of the function}
\item{status}{execution status}
\item{nonzero}{indices of initial observations with count > 0}
\item{dred}{result of \link{cdadadr} applied to non-zero observations}
\item{coreset}{indices of non-zero observations selected as core subset}
\item{clusters}{result of \link{QuickShiftClustering} applied to the core subset}
\item{theta}{distribution parameters for each cluster in the core subset}
\item{log2counts}{dithered and log2 transformed counts}
}
\description{
Background Read Density
}
\seealso{
\link{CPSES},
  \link{QuickShiftClustering},
  \link{cdadadr}
}
