% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdadadr.R
\name{cdadadr}
\alias{cdadadr}
\title{Count Density After Dithering And Dimensionality Reduction}
\usage{
cdadadr(cnt, knn, smobs = F, cvt = 0.5, npc = NA, zscore = T,
  dither = 1, progress = F)
}
\arguments{
\item{cnt}{matrix of read counts
(rows = observations, columns = measurement conditions).}

\item{knn}{number of nearest neighbors, which corresponds to the smoothing parameter
of estimated densities (larger values = smoother).}

\item{smobs}{subtract the mean count value of each observation (logical, default = F).}

\item{cvt}{percentage of variance to be ignored.
The value of \code{cvt} determines the stringency of the dimensionality
reduction which is performed by principal component analysis using
\link{prcomp}.
Dimensions retained after projection of the read count dataset on its
principal components will preserve \code{100 * (1 - cvt)} percents
of the initial variance.
For instance, after dimensionality reduction with the default setting
(\code{cvt} = 0.5), approximately 50 percents of the initial variance
will remain in the reduced data.
Setting \code{cvt} = 0 bypasses the dimensionality reduction and performs
density estimations in the original multivariate measurement space
of the dataset.}

\item{npc}{number of dimensions retained after projection on principal components.
When \code{npc} is specified, the \code{cvt} parameter is ignored.}

\item{zscore}{transform counts into z-scores (logical, default = T).}

\item{dither}{number of replicates for count dithering:
1 = single (default), 2 = duplicate, 3 = triplicate, etc.}

\item{progress}{show progress bar (logical, default = F).}
}
\value{
cdadadr returns a \code{list} with the following elements:
\item{parameters}{list with the value of each cdadadr argument}
\item{status}{list with informations on the dimensionality reduction}
\item{density}{knn density}
\item{projection}{transformed count matrix}
}
\description{
Count Density After Dithering And Dimensionality Reduction
}
\seealso{
\link{BRD},
  \link{DitherCounts},
  \link{knn_density}
}
\keyword{internal}
