% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_stats.R
\name{cdadadr}
\alias{cdadadr}
\title{Count Density After Dithering And Dimensionality Reduction}
\usage{
cdadadr(cnt, k, dither = 1, dred = 0.05, smobs = F, zscore = F)
}
\arguments{
\item{cnt}{matrix of read counts (rows = observations, columns = samples or conditions).}

\item{k}{number of nearest neighbors which corresponds to the smoothing parameter
of estimated densities (larger values = smoother).}

\item{dither}{number of replicates of the dithered counts: 0 = no dithering,
1 = single (default), 2 = duplicate, 3 = triplicate, etc.}

\item{dred}{stringency of the dimensionality reduction which is performed by
Principal Component Analysis. Dimensions retained after projection of the
read count dataset on its principal components will preserve
100 * (1 - dred) % of the total variance.
For instance with the default value (dred = 0.05), 95% of the total variance
will be retained after dimensionality reduction.
Setting dred = 0 bypasses the dimensionality reduction and performs
density estimations in the original multivariate space of the dataset.}

\item{smobs}{subtract the mean count value of each observation}

\item{zscore}{transform counts into z-scores (logical, default = F)}
}
\value{
cdadadr returns a \code{list} with the following elements:
\item{density}{knn density}
\item{parameters}{list with the value of each cdadadr argument}
\item{projection}{list}
}
\description{
Count Density After Dithering And Dimensionality Reduction
}
\seealso{
\link{DitherCounts},
  \link{prcomp},
  \link{knn_density}
}
