% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicRanges.R
\name{CleanupGRanges}
\alias{CleanupGRanges}
\title{Cleanup genomic ranges}
\usage{
CleanupGRanges(grg, seqinfo, organism, blacklist = NULL,
  keep.strand = FALSE)
}
\arguments{
\item{grg}{a \link{GRanges} object.}

\item{seqinfo}{\link{Seqinfo} object for the considered organism.}

\item{organism}{name of the organism (e.g. "Homo sapiens", "Mus musculus").}

\item{blacklist}{an optional \link{GRanges} object used to filter out any genomic ranges
overlaping with the blacklisted one (default = NULL, none).}

\item{keep.strand}{logical, when TRUE strand information is retained, when FALSE (default)
strand information is discarded (i.e. strand = "*").}
}
\value{
\code{CleanupGRanges} returns a \link{GRanges} object.
}
\description{
Cleanup genomic ranges such as imported ChIP-seq peaks from bed files by
removing ranges corresponding to non canonical chromosomes.
}
\seealso{
\link{ImportGenomicRanges},
  \link{ImportCpGIslandExt}
}
