% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{diaz.SES}
\alias{diaz.SES}
\title{Signal Extraction Scaling}
\usage{
diaz.SES(x, y)
}
\arguments{
\item{x}{Input or IgG raw read counts}

\item{y}{ChIP raw read counts}
}
\value{
list
}
\description{
Signal Extraction Scaling
}

