% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPSES.R
\name{Diaz.SES}
\alias{Diaz.SES}
\title{Signal Extraction Scaling (Diaz et al., 2012)}
\usage{
Diaz.SES(x, y, o = NULL)
}
\arguments{
\item{x}{vector of read counts in the control condition (Input, IgG, etc.)}

\item{y}{vector of read counts in the specific condition.}

\item{o}{precomputed ordering of \code{y} (default = NULL).}
}
\value{
\code{Diaz.SES} returns a scaling factor.
}
\description{
Signal Extraction Scaling (Diaz et al., 2012)
}
\seealso{
\link{CPSES}
}
\keyword{internal}
