% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{DitherCounts}
\alias{DitherCounts}
\title{Apply dithering to read counts}
\usage{
DitherCounts(x)
}
\arguments{
\item{x}{matrix of read counts (rows = observations, columns = samples or conditions).}
}
\value{
\code{DitherCounts} returns a matrix of dithered counts.
}
\description{
The \code{DitherCounts} function applies a triangular dithering filter
which eliminates the presence of identical values in a read count matrix.
}
\seealso{
\link{NonZeroCounts},
  \link{ReadCountMatrix}
}
