% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{DitherCounts}
\alias{DitherCounts}
\title{Dither counts}
\usage{
DitherCounts(x)
}
\arguments{
\item{x}{matrix of read counts (rows = observations, columns = samples or conditions).}
}
\value{
DitherCounts returns a \code{matrix} of dithered counts
}
\description{
Dither counts
}
