% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{ditherCounts}
\alias{ditherCounts}
\title{Dither counts}
\usage{
ditherCounts(x)
}
\arguments{
\item{x}{matrix of read counts (columns = samples, rows = intervals)}
}
\value{
ditherCounts returns a \code{matrix} of dithered counts
}
\description{
Dither counts
}
