% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{DitherCounts}
\alias{DitherCounts}
\title{dithering of read counts}
\usage{
DitherCounts(x)
}
\arguments{
\item{x}{matrix of read counts (rows = observations, columns = samples or conditions).}
}
\value{
\code{DitherCounts} returns a matrix of dithered counts.
}
\description{
dithering of read counts
}
\seealso{
\link{NonZeroCounts},
\link{MakeReadCountMatrix}
}
