% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{DitherCounts}
\alias{DitherCounts}
\title{Apply dithering to read counts}
\usage{
DitherCounts(x)
}
\arguments{
\item{x}{matrix of read counts (rows = observations, columns = samples or conditions).}
}
\value{
\code{DitherCounts} returns a matrix of dithered counts.
}
\description{
Apply dithering to read counts
}
\seealso{
\link{NonZeroCounts},
  \link{ReadCountMatrix},
  \link{MakeReadCounts}
}
