% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_matrixes.R
\name{ExtractColumns}
\alias{ExtractColumns}
\title{Extract count data}
\usage{
ExtractColumns(x, lst)
}
\arguments{
\item{x}{list of read count matrices
(rows = observations, columns = measurement conditions).}

\item{lst}{names or indices of columns to be extracted.}
}
\value{
\code{ExtractColumns} returns a list of read count matrices.
}
\description{
Extract count data
}
\seealso{
\link{JoinColumns},
\link{RemoveColumns},
\link{ReadCountMatrix},
\link{MakeReadCounts}
}
\keyword{internal}
