% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{FiniteValues}
\alias{FiniteValues}
\title{Detect computable values (i.e. not NA nor Inf)}
\usage{
FiniteValues(x)
}
\arguments{
\item{x}{numeric vector or matrix}
}
\value{
FiniteValues returns a \code{logical} vector
}
\description{
Detect computable values (i.e. not NA nor Inf)
}
\keyword{internal}
