% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{FiniteValues}
\alias{FiniteValues}
\title{Localise safe numeric values (i.e. not NA nor Inf)}
\usage{
FiniteValues(x)
}
\arguments{
\item{x}{numeric vector or matrix.}
}
\value{
\code{FiniteValues} returns a logical vector.
}
\description{
Localise safe numeric values (i.e. not NA nor Inf)
}
\seealso{
\link{NonZeroCounts},
\link{DetectCounts}
}
\keyword{internal}
