% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicRanges.R
\name{GenomicTiling}
\alias{GenomicTiling}
\title{Whole genome tiling with regular bins (i.e. running windows)}
\usage{
GenomicTiling(g, s, w = NULL)
}
\arguments{
\item{g}{\link{Seqinfo} object for the considered organism.}

\item{s}{integer, spacing between genomic bins in bp.}

\item{w}{integer, width of genomic bins in bp.}
}
\value{
\code{GenomicTiling} returns a \link{GRanges} object.
}
\description{
Whole genome tiling with regular bins (i.e. running windows)
}
\seealso{
\link{ImportGenomicRanges},
  \link{ImportCpGIslandExt},
  \link{BuildGeneFeatures},
  \link{ReadCountMatrix}
}
