% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCSC.R
\name{ImportCpGIslandExt}
\alias{ImportCpGIslandExt}
\title{Import CpG-islands from UCSC genomes}
\usage{
ImportCpGIslandExt(genome = NULL, fpath = NULL,
  goldenPath = "http://hgdownload.soe.ucsc.edu/goldenPath",
  cpgIslandExt = "database/cpgIslandExt.txt.gz", keep.file = FALSE, ...)
}
\arguments{
\item{genome}{name of an UCSC genome build. When \code{genome} is unspecified, the
\code{fpath} argument must be specified.}

\item{fpath}{path to a cpgIslandExt.txt.gz file. When \code{fpath} is unspecified, the
\code{genome} argument must be specified.}

\item{goldenPath}{base URL of UCSC genome repositories
(default = "http://hgdownload.soe.ucsc.edu/goldenPath").}

\item{cpgIslandExt}{path to cpgIslandExt.txt.gz files in the UCSC genome repositories
(default = "database/cpgIslandExt.txt.gz").}

\item{keep.file}{logical value controlling whether the downloaded file should be kept or
deleted (default = FALSE, delete).}

\item{...}{optional parameters forwarded to the \link{ImportGenomicRanges} function.}
}
\value{
\code{ImportCpGIslandExt} returns a \link{GRanges} object.
}
\description{
Import CpG-islands from UCSC genomes
}
\seealso{
\link{ImportGenomicRanges},
  \link{CleanupGRanges},
  \link{ReadCountMatrix}
}
