% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_matrixes.R
\name{JoinColumns}
\alias{JoinColumns}
\title{Merge count data}
\usage{
JoinColumns(x, y)
}
\arguments{
\item{x}{list of read count matrices
(rows = observations, columns = measurement conditions).}

\item{y}{list of read count matrices.}
}
\value{
\code{JoinColumns} returns a list of read count matrices.
}
\description{
Merge count data
}
\seealso{
\link{ExtractColumns},
\link{RemoveColumns},
\link{ReadCountMatrix},
\link{MakeReadCounts}
}
\keyword{internal}
