% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_stats.R
\name{knn_density}
\alias{knn_density}
\title{knn density estimator}
\usage{
knn_density(x, k, i = NULL, d = NULL, smoothing = T)
}
\arguments{
\item{x}{numeric matrix representing multivariate data where rows = observations
and columns = measurement conditions.}

\item{k}{number of nearest neighbors, which corresponds to the smoothing parameter
of estimated densities (larger values = smoother).}

\item{i}{precomputed matrix of nearest neighbor indices (optional).}

\item{d}{precomputed matrix of nearest neighbor distances (optional).}

\item{smoothing}{perfom a knn average smoothing of the density (default = T, recommended).}
}
\value{
knn_density returns a numeric vector.
}
\description{
k-nearest neighbor (knn) estimator of the density \deqn{P(Xi) ~ k / N Vi}
where \eqn{N} is the number of observations and \eqn{Vi} the volume of a
sphere of radius equal to the distance between \eqn{Xi} and its k-nearest
neighbor.
}
\seealso{
\link{knn_musigma2},
  \link{knn_mean}
}
