% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_stats.R
\name{knn_density}
\alias{knn_density}
\title{k-nearest neighbors estimator of density P(x) ~ k / N V}
\usage{
knn_density(x, nn.idx, nn.dist, k)
}
\arguments{
\item{x}{numeric matrix representing multivariate observations}

\item{nn.idx}{matrix of nearest neighbor indexes}

\item{nn.dist}{matrix of nearest neighbor distances}

\item{k}{number of nearest neighbors}
}
\value{
knn_density returns a vector of densities
}
\description{
k-nearest neighbors estimator of density P(x) ~ k / N V
}
