% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_stats.R
\name{knn_musigma2}
\alias{knn_musigma2}
\title{Local mean and variance}
\usage{
knn_musigma2(x, k, i = NULL, d = NULL, smoothing = F)
}
\arguments{
\item{x}{numeric matrix representing multivariate data where rows = observations
and columns = measurement conditions.}

\item{k}{number of nearest neighbors (i.e. smoothing factor).}

\item{i}{precomputed matrix of nearest neighbor indices (optional).}

\item{d}{precomputed matrix of nearest neighbor distances (optional).}

\item{smoothing}{perfom a knn average smoothing of the standard deviations (default = F).}
}
\value{
knn_musigma2 returns a \code{list} with the following elements:
\item{mu}{knn centroids}
\item{sigma2}{knn variances (squared standard deviations)}
}
\description{
k-nearest neighbor (knn) centroid and corresponding standard deviation.
}
\seealso{
\link{knn_density}
}
