% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_stats.R
\name{knn_musigma2}
\alias{knn_musigma2}
\title{Local mean and variance}
\usage{
knn_musigma2(x, k, i = NULL, d = NULL, smoothing = F)
}
\arguments{
\item{x}{numeric matrix representing multivariate data where rows = observations
and columns = measurement conditions.}

\item{k}{number of nearest neighbors, which corresponds to the smoothing parameter
of estimated densities (larger k values = smoother).}

\item{i}{precomputed matrix of nearest neighbor indices (optional).}

\item{d}{precomputed matrix of nearest neighbor distances (optional).}

\item{smoothing}{perfom a local average smoothing of
the estimated density for \link{knn_density} (default = T)
or the local variance for \link{knn_musigma2} (default = F).}
}
\value{
\code{knn_musigma2} returns a list with the following elements:
\item{mu}{knn centroids}
\item{sigma2}{knn variances (squared standard deviations)}
}
\description{
k-nearest neighbor (knn) centroid and corresponding standard deviation.
}
\seealso{
\link{knn_density},
  \link{knn_mean}
}
\keyword{internal}
