% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_stats.R
\name{knn_values}
\alias{knn_values}
\title{Extract local values based on a knn index matrix}
\usage{
knn_values(v, i)
}
\arguments{
\item{v}{numeric vector.}

\item{i}{precomputed matrix of nearest neighbor indices.}
}
\value{
\code{knn_values} returns a matrix.
}
\description{
Extract local values based on a knn index matrix
}
\seealso{
\link{knn_density},
  \link{knn_musigma2},
  \link{knn_mean}
}
\keyword{internal}
