% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{lc2ma}
\alias{lc2ma}
\title{Converts log2(counts) to AM values}
\usage{
lc2ma(x, y = NULL)
}
\arguments{
\item{x}{numeric vector, or matrix or list}

\item{y}{numeric vector}
}
\value{
\code{list} with a and m values
}
\description{
a = (y + x)/2, m = (y - x)
}
\seealso{
\link{ma2lc}
}
\keyword{internal}
