% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_matrixes.R
\name{LoadMappedReads}
\alias{LoadMappedReads}
\title{load/verify a list of mapped read files}
\usage{
LoadMappedReads(aln, NameByFile = F, ...)
}
\arguments{
\item{aln}{a vector of bam file paths or a list with named elements and containing
exclusively \link{GAlignments} objects.}

\item{NameByFile}{use file names to generate the name of each measurement condition, meaning
each sequenced biological sample (logical, default = F).}

\item{...}{additional arguments passed to the \link{readGAlignments} function.}
}
\value{
\code{LoadMappedReads} returns a list of \link{GAlignments} objects.
}
\description{
load/verify a list of mapped read files
}
\seealso{
\link{MakeReadCounts},
\link{ReadCountMatrix},
\link{AppendReadCounts},
\link{UpdateReadCounts}
}
\keyword{internal}
