% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{makeReadCountMatrix}
\alias{makeReadCountMatrix}
\title{Count mapped reads overlapping with genomic intervals}
\usage{
makeReadCountMatrix(aln, grg, ...)
}
\arguments{
\item{\code{aln}}{list of \link{GAlignments} objects with mapped reads from different samples}

\item{\code{grg}}{\link{GRanges} object defining the considered genomic intervals}
}
\value{
\code{matrix} of read counts (columns = samples, rows = genomic intervals)
}
\description{
Count mapped reads overlapping with genomic intervals
}

