% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadCountSimulation.R
\name{MakeSimulation}
\alias{MakeSimulation}
\title{Simulate ChIP-seq read count matrixes}
\usage{
MakeSimulation(p = NULL, m = NULL, f = NULL)
}
\arguments{
\item{p}{integer vector, population sizes}

\item{m}{matrix of parameters produced by the \link{DefineSimulation} function.}

\item{f}{list of functions allowing to use custom generators for simulated
measurements (default = NULL, recommended).}
}
\value{
\code{MakeSimulation} returns a \code{list} with the following elements:
\item{data}{matrix of simulated ChIP-seq read counts.}
\item{group}{population memberships.}
}
\description{
Simulate ChIP-seq read count matrixes
}
\seealso{
\link{DefineSimulation}
}
