% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_gmm.R
\name{mv_init_param}
\alias{mv_init_param}
\title{Multivariate normal distribution: parameter initialization}
\usage{
mv_init_param(x, ns)
}
\arguments{
\item{x}{numeric matrix representing multivariate data where rows = observations
and columns = samples or conditions.}

\item{ns}{number of normal distributions to be initialised}
}
\value{
\code{mv_init_param} returns a list of normal distribution parameters.
}
\description{
Multivariate normal distribution: parameter initialization
}
\seealso{
\link{mv_gmm},
\link{mv_emloop}
}
\keyword{internal}
