% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvfit.R
\name{mv_mle}
\alias{mv_mle}
\title{Multivariate normal distribution: maximum likelihood estimation}
\usage{
mv_mle(x, w = NULL, ML = T)
}
\arguments{
\item{x}{numeric matrix representing multivariate observations}

\item{w}{numeric weight (i.e. probability) associated to observations}

\item{ML}{logical (reserved, default = T)}
}
\value{
mv_mle returns a \code{list} with the following elements:
\item{mu}{mean vector}
\item{sigma}{covariance matrix}
}
\description{
Multivariate normal distribution: maximum likelihood estimation
}
