% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_gmm.R
\name{mv_mle}
\alias{mv_mle}
\title{Multivariate normal distribution: maximum likelihood estimation}
\usage{
mv_mle(x, w = NULL, ML = T)
}
\arguments{
\item{x}{numeric matrix representing multivariate data where rows = observations
and columns = samples or conditions.}

\item{w}{numeric weight (i.e. probability) associated to observations}

\item{ML}{logical (reserved, default = T)}
}
\value{
\code{mv_mle} returns a list with the following elements:
\item{mu}{mean vector}
\item{sigma}{covariance matrix}
}
\description{
Multivariate normal distribution: maximum likelihood estimation
}
\seealso{
\link{mv_pdf},
\link{mv_rsg},
\link{mv_gmm}
}
\keyword{internal}
