% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_gmm.R
\name{mv_rsg}
\alias{mv_rsg}
\title{Multivariate normal distribution: random sample generator}
\usage{
mv_rsg(n, theta)
}
\arguments{
\item{n}{number of observations}

\item{theta}{list of normal distribution parameters}
}
\value{
mv_rsg returns a numeric matrix representing multivariate observations
}
\description{
Multivariate normal distribution: random sample generator
}
\keyword{internal}
