% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_samples}
\alias{plot_samples}
\title{plot_samples}
\usage{
plot_samples(x, idx = 1:2, prm = NULL, grp = NULL, grp_order = NULL,
  symetric = T, clr = NULL, alpha = 0.1, ...)
}
\arguments{
\item{x}{matrix}

\item{idx}{vector}

\item{prm}{data.frame}

\item{grp}{vector}

\item{grp_order}{vector}

\item{symetric}{logical}

\item{clr}{charcater}

\item{alpha}{numeric}
}
\description{
plot_samples
}
\keyword{internal}
