% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBRD.R
\name{PlotBRD}
\alias{PlotBRD}
\title{PlotBRD}
\usage{
PlotBRD(brd, core.only = F, with.legend = T)
}
\arguments{
\item{brd}{result from \link{BRD}}

\item{core.only}{logical (default = F)}

\item{with.legend}{logical (default = T)}
}
\description{
PlotBRD
}
\keyword{internal}
