% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBRD.R
\name{PlotBRD}
\alias{PlotBRD}
\title{Control plots from BRD results}
\usage{
PlotBRD(brd, with.axes = T, with.legend = T, title = "", res = 400)
}
\arguments{
\item{brd}{result from the \link{BRD} function.}

\item{with.axes}{logical, include plot axes (default = T).}

\item{with.legend}{logical, include legends (default = T).}

\item{title}{a title for generated plots (optional).}

\item{res}{number of bins used to make the intensity-density scatterplot
(default = 400).}
}
\description{
Control plots from BRD results
}
\seealso{
\link{BRD},
  \link{CDaDaDR}
}
