% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBRD_V3.R
\name{PlotBRD}
\alias{PlotBRD}
\title{Control plots from BRD results}
\usage{
PlotBRD(brd, title = "", plots = c("density", "clusters", "thresholds",
  "distributions"), palette = "magma", gradient = "hcl.duo.light",
  saturation = 1, with.axes = TRUE, with.legend = TRUE, res = 400)
}
\arguments{
\item{brd}{result from calling the \link{BRD} function.}

\item{title}{common title for the generated plots (default = none).}

\item{plots}{character vector specifying which plots are being generated. By default
\code{PlotBRD} generates 4 plots named "density", "clusters", "thresholds",
and "distributions". It is possible to generate only a subset of these plots
by using the corresponding vector of plot names, like for instance
\code{plots = c("density", "thresholds")}.}

\item{with.axes}{logical, include plot axes or not (default = TRUE, yes).}

\item{with.legend}{logical, include legends or not (default = TRUE, yes).}

\item{res}{resolution in number of bins for the "thresholds" plot (default = 400).}
}
\description{
The \code{PlotBRD} function produces control graphs representing intermediate
results of the \link{BRD} function. These control graphs are useful to adapt
parameters of the BRD method to the considered read count matrix.
}
\seealso{
\link{BRD},
  \link{CDaDaDR.2D}
}
