% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickshift.R
\name{QuickShift}
\alias{QuickShift}
\title{QuickShift algorithm (Vedaldi & Soatto, 2008)}
\usage{
QuickShift(x, d, progress = F)
}
\arguments{
\item{x}{numeric matrix representing multivariate data where rows = observations
and columns = samples or conditions.}

\item{d}{numeric vector representing the density estimation at each observation.}

\item{progress}{progress bar (logical, default = F)}
}
\value{
QuickShift returns a graph object (igraph package)
}
\description{
QuickShift algorithm (Vedaldi & Soatto, 2008)
}
\keyword{internal}
