% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickshift.R
\name{QuickShiftClustering}
\alias{QuickShiftClustering}
\title{hierarchical clustering algorithm based on density gradient ascent}
\usage{
QuickShiftClustering(x, d, n, ...)
}
\arguments{
\item{n}{number of clusters.}

\item{...}{}
}
\value{
\code{QuickShiftClustering} returns a list with the following elements:
\item{membership}{vector of cluster memberships}
\item{csize}{cluster sizes}
\item{center}{indice of the maximum density value in each cluster}
}
\description{
hierarchical clustering algorithm based on density gradient ascent
}
\seealso{
\link{QuickShift},
\link{QuickShiftCutClusters}
}
