% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickshift.R
\name{QuickShiftClustering}
\alias{QuickShiftClustering}
\title{Hierarchical clustering based on density gradient ascent}
\usage{
QuickShiftClustering(x, d, n, ...)
}
\arguments{
\item{x}{numeric matrix representing multivariate data where rows = observations
and columns = measurement conditions.}

\item{d}{numeric vector representing a density estimation at each observation.}

\item{n}{desired number of clusters.}

\item{...}{}
}
\value{
\code{QuickShiftClustering} and \code{QuickShiftClusters} return a list
with the same following elements:
\item{membership}{
  vector of integers in [1, \code{n}] indicating to which cluster each
  observation belongs.
}
\item{sizes}{number of observations in each cluster.}
}
\description{
Hierarchical clustering based on density gradient ascent
}
\references{
Vedaldi A., Soatto S. (2008) Quick Shift and Kernel Methods for Mode Seeking.
In: Forsyth D., Torr P., Zisserman A. (eds) Computer Vision – ECCV 2008.
ECCV 2008. Lecture Notes in Computer Science, vol 5305.
Springer, Berlin, Heidelberg
\url{http://dx.doi.org/10.1007/978-3-540-88693-8_52}
}
\seealso{
\link{QuickShift},
  \link{QuickShiftClusters}
}
\keyword{internal}
