% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickshift.R
\name{QuickShiftCutClusters}
\alias{QuickShiftCutClusters}
\title{QuickShiftCutClusters}
\usage{
QuickShiftCutClusters(g, n = NULL, ecut = NULL)
}
\arguments{
\item{g}{QuickShift graph resulting from the \link{QuickShift} function}

\item{n}{expected number of clusters}

\item{ecut}{maximum branch length}
}
\value{
QuickShiftCutClusters returnd a \code{list} with the following elements
}
\description{
QuickShiftCutClusters
}
\seealso{
\link{QuickShiftClustering}
}
\keyword{internal}
