% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickshift.R
\name{quickShiftCutClusters}
\alias{quickShiftCutClusters}
\title{quickshift graph partition}
\usage{
quickShiftCutClusters(g, n = NULL, ecut = NULL)
}
\arguments{
\item{g}{quickshift graph resulting from the \link{quickShift} function}

\item{n}{expected number of clusters}

\item{ecut}{maximum branch length}
}
\value{
quickShiftCutClusters return a \code{list} with the following elements
}
\description{
quickshift graph partition
}
