% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickshift.R
\name{quickShiftSearchClusters}
\alias{quickShiftSearchClusters}
\title{quickshift root and top level branches}
\usage{
quickShiftSearchClusters(i = 0, d = 0, dmax)
}
\arguments{
\item{i}{node index}

\item{d}{depth level}

\item{dmax}{maximum depth}
}
\value{
quickShiftSearchClusters returns a \code{data.frame}
}
\description{
quickshift root and top level branches
}
