% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_matrixes.R
\name{ReadCountMatrix}
\alias{ReadCountMatrix}
\title{count reads in genomic intervals for different conditions}
\usage{
ReadCountMatrix(aln, grg, cnt = NULL, update = F, NameByFile = F,
  maxgap = 0L, minoverlap = 1L, type = "any", ignore.strand = T, ...)
}
\arguments{
\item{aln}{a vector of bam file paths or a list with named elements and containing
exclusively \link{GAlignments} objects.}

\item{grg}{a \link{GRanges} object defining the genomic intervals of interest.}

\item{cnt}{a numeric matrix representing read counts,
where rows = observations and columns = measurement conditions.}

\item{update}{instead of producing new count matrixes from scratch, this argument allows
to update targeted count columns when existing count matrixes are provided
(logical, default = F).}

\item{NameByFile}{use file names to generate the name of each measurement condition, meaning
each sequenced biological sample (logical, default = F).}

\item{maxgap}{
    See \code{\link[IRanges]{findOverlaps}} in the \pkg{IRanges} package
    for a description of these arguments.
  }

\item{minoverlap}{
    See \code{\link[IRanges]{findOverlaps}} in the \pkg{IRanges} package
    for a description of these arguments.
  }

\item{type}{
    See \code{\link[IRanges]{findOverlaps}} in the \pkg{IRanges} package
    for a description of these arguments.
  }

\item{ignore.strand}{
    When set to \code{TRUE}, the strand information is ignored in the
    overlap calculations.
  }

\item{...}{additional arguments passed to the \link{readGAlignments} function.}
}
\value{
functions \code{ReadCountMatrix}, \code{AppendReadCounts}, and
\code{UpdateReadCounts} return a read counts matrix where
rows = observations and columns = measurement conditions.
}
\description{
\code{ReadCountMatrix} computes the number of reads overlapping with provided
genomic intervals for different measurement conditions.
}
\seealso{
\link{LoadMappedReads},
\link{AppendReadCounts},
\link{UpdateReadCounts},
\link{NonZeroCounts},
\link{DitherCounts}
}
