% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{scalingFactors}
\alias{scalingFactors}
\title{Scaling factors between read counts in different samples}
\usage{
scalingFactors(cnt, rpkb, background = NULL, enriched = NULL,
  replicates = 0, center = NULL, pairwise.density = T, d = 80,
  k = 128, plot = F, as.png = T, xylim = NULL, xylim.rpkb = NULL, ...)
}
\arguments{
\item{cnt}{matrix of log transformed read counts
(columns = samples, rows = genomic intervals)}

\item{rpkb}{matrix of log transformed reads per kb associated to the \code{cnt} matrix
(columns = samples, rows = genomic intervals)}

\item{background}{logical vector defining the background subset of counts (default = all)}

\item{enriched}{logical vector defining the enriched subset of counts (default = none)}

\item{replicates}{logical matrix indicating replicate samples (default = none)}

\item{center}{function for center of difference calculations.
The default center function is \link{findMode} which identifies the mode in
a kernel density estimation of the distribution of read count differences
using the generic R function \code{density}.
This default function can be replaced by \code{median}, \code{mean}, or
similar functions.}

\item{pairwise.density}{logical (default = T, recommended).
If \code{TRUE} count density estimations and subsequent selections of
counts assumed to exhibit minimal variations are performed in bi-variate
spaces corresponding to all pairwise sample comparisons.
If \code{FALSE} count density estimation is performed only once, in the
multi-variate space of all samples.}

\item{d}{density threshold (in percentage) used to select a subset of counts assumed
to reflect minimal variations}

\item{k}{number of nearest neighbors for kNN density estimation}

\item{...}{optional parameters passed to the \link{scalingFactors} function.
These include:}
}
\value{
scalingFactors returns a \code{list} with the following elements:
\item{pairwise.factors}{
  matrix of pairwise normalization factors between samples
  (\code{cnt} and \code{rpkb} columns)
}
\item{low.variation}{
  logical vector indicating the subset of rows assumed invariable by the
  normalization
}
}
\description{
Scaling factors between read counts in different samples
}

