% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRD_V3.R
\name{ScalingFactors}
\alias{ScalingFactors}
\title{Get BRD scaling factors}
\usage{
ScalingFactors(brd, as.log2 = FALSE)
}
\arguments{
\item{brd}{result of a prior call to the \link{BRD} function.}

\item{as.log2}{logical (default = FALSE, no).}
}
\value{
\code{ScalingFactors} returns a numeric vector.
}
\description{
Extract the vector of scaling factors from the result of a prior call to the
\link{BRD} function.
}
\seealso{
\link{BRD},
  \link{NormalizeCountMatrix}
}
