% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{Wave2D}
\alias{Wave2D}
\title{Wave2D}
\usage{
Wave2D(x, mu = NULL, f = NULL, r = NULL, p = 0, e = 1, a = -1)
}
\arguments{
\item{x}{matrix of sampling coordinates in the 2D space.}

\item{mu}{matrix of source location coordinates in the 2D space.}

\item{f}{frequencies vector}

\item{r}{radius matrix}

\item{p}{phases vector}

\item{e}{energy vector,}

\item{a}{attenuation vector}
}
\value{
Wave2D returns a vector
}
\description{
generator of 2D waves \deqn{(u + e)^a * cos(p + 2 * pi * u * f)}
where \eqn{u} is the distance to the source.
}
\keyword{internal}
