% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{xy2md}
\alias{xy2md}
\title{Convert (x, y) to (a = mean, m = difference) coordinates}
\usage{
xy2md(x, y = NULL)
}
\arguments{
\item{x}{numeric vector, or a matrix with 2 columns or a list containing two vectors.}

\item{y}{numeric vector.}
}
\value{
\code{xy2md} returns a list or matrix with (a = mean, m = difference) values.
}
\description{
compute \eqn{a = (y + x)/2}, and \eqn{m = (y - x)}
The inverse transformation of \code{xy2md} is \link{md2xy}.
}
\seealso{
\link{md2xy}
}
\keyword{internal}
